% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cyclone_energy.R
\name{calculate_cyclone_energy}
\alias{calculate_cyclone_energy}
\title{Calculate the Accumulated Cyclone Energy of a Storm}
\usage{
calculate_cyclone_energy(example_infile)
}
\arguments{
\item{example_infile}{Describe Input}
}
\value{
Describe Output
}
\description{
This function computes the accumulated cyclone energy of a given storm.
}
