% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{computeExposure}
\alias{computeExposure}
\title{rasterizeExposure counterpart function for non raster data}
\usage{
computeExposure(wind)
}
\arguments{
\item{wind}{numeric vector. Wind speed values}
}
\value{
numeric vector of length 5 (for each category).
Exposure computed using the wind speed values in wind
}
\description{
rasterizeExposure counterpart function for non raster data
}
