% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{computePDI}
\alias{computePDI}
\title{rasterizePDI counterpart function for non raster data}
\usage{
computePDI(wind)
}
\arguments{
\item{wind}{numeric vector. Wind speed values}
}
\value{
numeric. PDI computed using the wind speed values in wind
}
\description{
rasterizePDI counterpart function for non raster data
}
