% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{computeProduct}
\alias{computeProduct}
\title{rasterizeProduct counterpart function for non raster data}
\usage{
computeProduct(product, wind, result)
}
\arguments{
\item{product}{character. Product input from stormBehaviour}

\item{wind}{numeric vector. Wind speed values}

\item{result}{numeric array. Similar as final_stack, i.e where to add the
computed product}
}
\value{
numeric array of dimension:
\itemize{
\item number of observations: number of points. If product == "MSW"
\item 1 : number of points. If product == "PDI"
\item 5 : number of points. If product == "Exposure"
}
}
\description{
rasterizeProduct counterpart function for non raster data
}
