% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTEW.R
\name{computeTEW.list}
\alias{computeTEW.list}
\title{Compute the topographic exposure to wind}
\usage{
\method{computeTEW}{list}(profiles, points, dtm, angle = 6, threshold = 0, verbose = 2, ...)
}
\arguments{
\item{profiles}{list Wind speed and direction profiles from \code{temporalBehaviour()} function which must respect
the following terminology : 'speed', 'direction' columns in a data.frame}

\item{points}{data.frame. Consisting of two columns names as "x" (for the longitude) and
"y" (for the latitude), providing the coordinates in decimal degrees of the point locations.
Same as the "points" argument of the \code{temporalBehaviour()} function}

\item{dtm}{SpatRaster of the elevation data (Digital Terrain Model) for a given location}

\item{angle}{numeric. Inflection angle of the wind (in degrees). default is 6°.}

\item{threshold}{numeric. Minimum wind speed threshold (in m/s) requirred to compute exposure. default is 0}

\item{verbose}{numeric. Whether or not the function should display
information about the process and/or outputs. Can be:
\itemize{
\item \code{2}, information about the processes and outputs are displayed (default setting),
\item \code{1}, information about the processes are displayed,
\item \code{0}, no information displayed.
}}

\item{...}{additional arguments (not used)}
}
\value{
the function returns a list of data.frame with one additional column for topographic exposure to wind (TEW)
for each observation or interpolated observation and each \code{Storm}.
}
\description{
The \code{computeTEW()} function allows computing topographic exposure to wind (TEW)
for each points registred in a data.frame for a tropical cyclone or set of tropical cyclones.
}
