% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTEW.R
\name{computeTEW.SpatRaster}
\alias{computeTEW.SpatRaster}
\title{Compute the topographic exposure to wind}
\usage{
\method{computeTEW}{SpatRaster}(
  profiles,
  sts,
  dtm,
  angle = 6,
  threshold = 0,
  product = "TEWIntegrated",
  verbose = 2,
  ...
)
}
\arguments{
\item{profiles}{SpatRaster. Wind speed and direction profiles from \code{spatialBehaviour()} function which must respect
the following terminology : 'STORM_Speed_N', 'STORM_Direction_N'}

\item{sts}{StormsList object}

\item{dtm}{SpatRaster of the elevation data (Digital Terrain Model) for a given location}

\item{angle}{numeric. Inflection angle of the wind (in degrees). default is 6°.}

\item{threshold}{numeric. Minimum wind speed threshold (in m/s) requirred to compute exposure. default is 0}

\item{product}{character vector. Desired output statistics:
\itemize{
\item \code{"TEW1wd"}, TEW computation at each observation, one wind direction taken where wind speed is maximal.
Good if your loi is small as wind is considered spatially homogeneous
\item \code{"TEW1wdMax"}, computes \code{"TEW1wd"}, then takes the maximum of the layers
\item \code{"TEW1wdMean"}, computes \code{"TEW1wd"}, then takes the mean of the layers
\item \code{"TEW"}, TEW computation at each observation, wind direction varies spatially
\item \code{"TEWIntegrated"}, TEW computation for maximum wind speed over the whole storm track for each spatial
location of the raster. Use wind direction associated with the maximal wind speed (default)
}}

\item{verbose}{numeric. Whether or not the function should display
information about the process and/or outputs. Can be:
\itemize{
\item \code{2}, information about the processes and outputs are displayed (default setting),
\item \code{1}, information about the processes are displayed,
\item \code{0}, no information displayed.
}}

\item{...}{additional arguments (not used)}
}
\value{
the function returns one layer for topographic exposure to wind (TEW)
for each observation or interpolated observation and each \code{Storm}.
The names of the layer follow the following terminology, the name of the storm in capital letters,
"TEW", and the indices of the observation separated by underscores
(e.g., "PAM_TEW_41", ...)
}
\description{
The \code{computeTEW()} function allows computing topographic exposure to wind (TEW)
for each cell of a raster grid for a tropical cyclone or set of tropical cyclones.
}
