% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eezNC}
\alias{eezNC}
\title{EEZ of New Caledonia}
\format{
\subsection{\code{eezNC}}{

a sf object
}
}
\usage{
eezNC
}
\description{
Provide the geographic limits of the eez of New Caledonia
}
\keyword{datasets}
