% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{get_RMW}
\alias{get_RMW}
\title{Get the radius of maximum wind speed according to Willoughby model}
\usage{
get_RMW(w_max, lat)
}
\arguments{
\item{w_max}{Maximum wind speed generated}

\item{lat}{Latitude where \code{w_max} has occured}
}
\value{
radius of maximum wind speed (km)
}
\description{
Get the radius of maximum wind speed according to Willoughby model
}
