% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getBuffer}
\alias{getBuffer}
\alias{getBuffer,StormsList-method}
\title{Get the extended LOI from a Storms object}
\usage{
getBuffer(sts)

\S4method{getBuffer}{StormsList}(sts)
}
\arguments{
\item{sts}{\code{StormsList} object}
}
\value{
sf object. Extended Location Of Interest for the given \code{StormsList}
object
}
\description{
Get the Extended Location Of Interest from a \code{StormsList} object
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")
buff <- getBuffer(sts)
}
}
