% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getBuffer}
\alias{getBuffer}
\title{getBuffer}
\usage{
getBuffer(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
sf object. Extented location of interest for a given Storms object
}
\description{
getBuffer
}
