% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getBufferSize}
\alias{getBufferSize}
\alias{getBufferSize,StormsList-method}
\title{Get buffer size from a Storms object}
\usage{
getBufferSize(sts)

\S4method{getBufferSize}{StormsList}(sts)
}
\arguments{
\item{sts}{\code{StormsList} object}
}
\value{
numeric. Buffer size (in km) used to generate the extended location
of interest for the given \code{StormsList} object
}
\description{
Get the buffer size for a \code{StormsList} object
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")
buffsize <- getBufferSize(sts)
}
}
