% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getBufferSize}
\alias{getBufferSize}
\title{getBufferSize}
\usage{
getBufferSize(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
numeric. Buffer size (in km) used to generate the extended LOI for a given Storms object
}
\description{
getBufferSize
}
