% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{getColors}
\alias{getColors}
\title{Get the Saffir Simpson Hurricane scale colors associated with a maximum
sustained wind speed}
\usage{
getColors(msw)
}
\arguments{
\item{msw}{numeric. Maximum Sustained Wind}
}
\value{
color associated with the observation
}
\description{
Get the Saffir Simpson Hurricane scale colors associated with a maximum
sustained wind speed
}
