% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defStormsDataset.R
\name{getDataFromCsvFile}
\alias{getDataFromCsvFile}
\title{Load database when filename is a CSV
(CF defStormsDataset for additional informations about parameters)}
\usage{
getDataFromCsvFile(
  filename,
  sep,
  fields,
  basin,
  seasons,
  unitConversion,
  notNamed,
  verbose
)
}
\arguments{
\item{filename}{character}

\item{sep}{character. separator for the csv file in input}

\item{fields}{named character vecor}

\item{basin}{character}

\item{seasons}{numeric vector}

\item{unitConversion}{named character vector}

\item{notNamed}{character for not named storms}

\item{verbose}{numeric}
}
\value{
list of arrays
}
\description{
Load database when filename is a CSV
(CF defStormsDataset for additional informations about parameters)
}
