% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getInObs}
\alias{getInObs}
\alias{getInObs,Storms-method}
\title{getInObs}
\usage{
getInObs(sts, name, season = NULL)

\S4method{getInObs}{Storms}(sts, name, season = NULL)
}
\arguments{
\item{sts}{Storms object}

\item{name}{character. Name of the storm to extract}

\item{season}{numeric. Cyclonic season of the storm to extract. Use only if several
storms in the Storms object share the same name. Default value is set de NULL}
}
\value{
numeric. Indices within the extented LOI for the given Storm object
}
\description{
getInObs
}
