% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getLOI}
\alias{getLOI}
\alias{getLOI,Storms-method}
\title{getLOI}
\usage{
getLOI(sts)

\S4method{getLOI}{Storms}(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
sf object. Location of interest for a given Storms object
}
\description{
getLOI
}
