% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNames}
\alias{getNames}
\alias{getNames,Storms-method}
\title{getNames}
\usage{
getNames(sts)

\S4method{getNames}{Storms}(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
character vector. Names of each storms provided by a given Storms object
}
\description{
getNames
}
