% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbObs}
\alias{getNbObs}
\alias{getNbObs,Storm-method}
\alias{getNbObs,StormsList-method}
\title{Get the number of observations available}
\usage{
getNbObs(s, ...)

\S4method{getNbObs}{Storm}(s)

\S4method{getNbObs}{StormsList}(s, name, season = NULL)
}
\arguments{
\item{s}{\code{Storm}/\code{StormsList}}

\item{...}{extra arguments for \code{StormsList}}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the \code{Storm} to extract. Used only
if several \code{Storm} in \code{s} share the same name. Default value is set to
\code{NULL}}
}
\value{
numeric. Number of observations available for the given \code{Storm}
}
\description{
Get the number of observations available for a given \code{Storm}
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")

## For storm Niran
getNbObs(getStorm(sts, name = "NIRAN"))

## Equivalent to
getNbObs(sts, name = "NIRAN")
}
}
