% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbStorms}
\alias{getNbStorms}
\alias{getNbStorms,StormsList-method}
\title{Get number of \code{Storm} in \code{StormsList}}
\usage{
getNbStorms(sts)

\S4method{getNbStorms}{StormsList}(sts)
}
\arguments{
\item{sts}{\code{StormsList} object}
}
\value{
numeric. Number of storms available in the given \code{StormsList} object
}
\description{
Get the number of \code{Storm} available in a \code{StormsList} object
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")
nb <- getNbStorms(sts)
}
}
