% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getNbStorms}
\alias{getNbStorms}
\title{getNbStorms}
\usage{
getNbStorms(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
numeric. Number of storms available in a given Storms object
}
\description{
getNbStorms
}
