% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getObs}
\alias{getObs}
\title{getObs}
\usage{
getObs(sts, name, season = NULL)
}
\arguments{
\item{sts}{Storms object}

\item{name}{character. Name of the storm to extract}

\item{season}{numeric. Cyclonic season of the storm to extract. Use only if several
storms in the Storms object share the same name. Default value is set de NULL}
}
\value{
data.frame. Observations of the given storm
}
\description{
getObs
}
