% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{getRmw}
\alias{getRmw}
\title{Compute the Radius of Maximum Wind. It is an empirical formula extracted from
Willoughby model}
\usage{
getRmw(msw, lat)
}
\arguments{
\item{msw}{numeric. Maximum Sustained Wind (m/s)}

\item{lat}{numeric. Should be between -60 and 60. Latitude of the eye of the storm}
}
\value{
Radius of Maximum Wind (km)
}
\description{
Compute the Radius of Maximum Wind. It is an empirical formula extracted from
Willoughby model
}
