% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSeasons}
\alias{getSeasons}
\alias{getSeasons,Storm-method}
\alias{getSeasons,StormsList-method}
\title{Get the cyclonic seasons of storm(s)}
\usage{
getSeasons(s)

\S4method{getSeasons}{Storm}(s)

\S4method{getSeasons}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm}/\code{StormsList} object}
}
\value{
numeric vector. Cyclonic seasons of each storms (or the selected one
through \code{name} input) provided by the given \code{StormsList} object or the cyclonic
season for the given \code{Storm} object
}
\description{
Get the cyclonic season of storms available in a \code{Storm}/\code{StormsList}
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")

## For storm Niran
getSeasons(getStorm(sts, name = "NIRAN"))

## For sts
getSeasons(sts)
}
}
