% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getSeasons}
\alias{getSeasons}
\title{getSeasons}
\usage{
getSeasons(sts, name = NULL)
}
\arguments{
\item{sts}{Storms object}

\item{name}{character. Name(s) of the storm(s). Default value is set to NULL}
}
\value{
numeric vector. Cyclonic seasons of each storms (or the selected one
through names input)provided by a given Storms object
}
\description{
getSeasons
}
