% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorm_inObs}
\alias{getStorm_inObs}
\title{getStorm_inObs}
\usage{
getStorm_inObs(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
numeric. Indices within the extented LOI for a given Storm object
}
\description{
getStorm_inObs
}
