% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorm_name}
\alias{getStorm_name}
\title{getStorm_name}
\usage{
getStorm_name(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
character. Name of the storm
}
\description{
Getter to access name of the storm
}
