% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getStorm_nbObs}
\alias{getStorm_nbObs}
\alias{getStorm_nbObs,Storm-method}
\title{getStorm_nbObs}
\usage{
getStorm_nbObs(st)

\S4method{getStorm_nbObs}{Storm}(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
numeric. Number of observations available within the Storm object
}
\description{
getStorm_nbObs
}
