% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorm_nbObs}
\alias{getStorm_nbObs}
\title{getStorm_nbObs}
\usage{
getStorm_nbObs(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
numeric. Number of observations available within the Storm object
}
\description{
getStorm_nbObs
}
