% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getStorm_obs}
\alias{getStorm_obs}
\alias{getStorm_obs,Storm-method}
\title{getStorm_obs}
\usage{
getStorm_obs(st)

\S4method{getStorm_obs}{Storm}(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
data.frame. Observations for a given Storm object
}
\description{
getStorm_obs
}
