% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorm_obs}
\alias{getStorm_obs}
\title{getStorm_obs}
\usage{
getStorm_obs(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
data.frame. Observations for a given Storm object
}
\description{
getStorm_obs
}
