% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorm_season}
\alias{getStorm_season}
\title{getStorm_season}
\usage{
getStorm_season(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
numeric. Cyclonic season of the storm
}
\description{
getStorm_season
}
