% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getStorm_sshs}
\alias{getStorm_sshs}
\alias{getStorm_sshs,Storm-method}
\title{getStorm_sshs}
\usage{
getStorm_sshs(st)

\S4method{getStorm_sshs}{Storm}(st)
}
\arguments{
\item{st}{Storm object}
}
\value{
numeric. Maximum catergory reached in the Saffir Simpson Hurricane Scale
}
\description{
getStorm_sshs
}
