% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTEW.R
\name{getTerrain}
\alias{getTerrain}
\title{Get terrain data}
\usage{
getTerrain(elevation)
}
\arguments{
\item{elevation}{SpatRaster with elevation data}
}
\value{
list of SpatRaster with slope and aspect, in radians
}
\description{
Get terrain data
}
