% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{Holland80_cyc_profil}
\alias{Holland80_cyc_profil}
\title{Compute wind values according to the Holland 80 model}
\usage{
Holland80_cyc_profil(r, rmw, pres, poci, lat, b = 1.3)
}
\arguments{
\item{r}{Distance to the center of the storm where the value must be computed}

\item{rmw}{Radius of maximum wind speed. Default value is set to NULL}

\item{pres}{Pressure at the center of the storm}

\item{poci}{Pressure at the Outermost Closed Isobar}

\item{lat}{Latitude of the eye of the storm}

\item{b}{Scaling factor}
}
\value{
wind value according to the Holland 80 at distance \code{r} to the
center of the storm
}
\description{
Compute wind values according to the Holland 80 model
}
