% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hurdat}
\alias{hurdat}
\title{North Atlantic Hurricane Data}
\format{
\subsection{\code{hurdat}}{

A data frame with 53,972 rows and 23 columns.
}
}
\source{
\url{https://www.aoml.noaa.gov/hrd/hurdat/hurdat2.html}
}
\usage{
data(hurdat)
}
\description{
This dataset contains information on North Atlantic hurricanes from 1851 to 2022.
The data is from the National Hurricane Center's HURDAT2 dataset.
The data is available in the \code{hurdat} dataset.
}
\keyword{datasets}
