% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IBTRACS_SP}
\alias{IBTRACS_SP}
\title{South Pacific storms database}
\format{
\subsection{\code{IBTRACS_SP}}{

a StormsDataset object
}
}
\usage{
IBTRACS_SP
}
\description{
Database extracted from the IBTrACS.SP.v04r00.nc file.
It provides all the tropical cyclones that occured in the SP Basin from the 1867 to nowadays
}
\keyword{datasets}
