% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormDatabase.R
\name{initDatabase}
\alias{initDatabase}
\title{Title}
\usage{
initDatabase(
  url =
    "https://www.ncei.noaa.gov/data/international-best-track-archive-for-climate-stewardship-ibtracs/v04r00/access/netcdf/IBTrACS.ALL.v04r00.nc",
  name = "IBTrACS.ALL.v04r00.nc",
  fields = c(names = "name", seasons = "season", isoTime = "iso_time", lon = "usa_lon",
    lat = "usa_lat", msw = "usa_wind", rmw = "usa_rmw", roci = "usa_roci", pressure =
    "usa_pres", poci = "usa_poci", sshs = "usa_sshs", numobs = "numobs")
)
}
\arguments{
\item{url}{...}

\item{name}{...}

\item{fields}{...}
}
\value{
...
}
\description{
Title
}
