% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_storm_track.R
\name{interpolate_storm_track}
\alias{interpolate_storm_track}
\title{Interpolate Storm Track}
\usage{
interpolate_storm_track(df)
}
\arguments{
\item{df}{Hurricane Dataset containing of a single storm track}
}
\value{
Updated Hurricane Dataset with 30 minute increments interpolated
}
\description{
This function interpolates a storm track to 30 minute increments.
}
