% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made_landfall_US.R
\name{made_landfall_US}
\alias{made_landfall_US}
\title{Determine if the Storm Made Landfall in the United States}
\usage{
made_landfall_US(example_infile)
}
\arguments{
\item{example_infile}{Describe Input}
}
\value{
Describe Output
}
\description{
A function for determining whether the storm made landfall in the
continental United States based on its latitude and longitude values.
}
