% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_pos_and_size.R
\name{plot_storm_pos_and_size}
\alias{plot_storm_pos_and_size}
\title{Plot Storm Position and Size}
\usage{
plot_storm_pos_and_size(example_infile)
}
\arguments{
\item{example_infile}{Describe Input}
}
\value{
Describe Output
}
\description{
A function to plot a map of the position and size of a given storm.
The 34, 50, and 64 knot extent variables.
}
