% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_track.R
\name{plot_storm_track}
\alias{plot_storm_track}
\title{Plot Storm Tracks}
\usage{
plot_storm_track(df_list)
}
\arguments{
\item{df_list}{A list of data frames containing storm track data. Each data frame tracks one storm.}
}
\value{
A map of storm tracks.
}
\description{
This function plots a map of storm tracks for a selection of storms.
The map includes country and US state boundaries.
}
