% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plot_track}
\alias{plot_track}
\title{Plot the track of a storm on an existing map}
\usage{
plot_track(storm)
}
\arguments{
\item{storm}{Storm object that we want to plot the track. Note that a map
should be previously plotted}
}
\description{
Plot the track of a storm on an existing map
}
