% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBehaviour.R
\name{plotBehaviour}
\alias{plotBehaviour}
\title{Plot a selected product (MSW, PDI ...) alongside with the track of a selected
storm contained in object \code{sts}.}
\usage{
plotBehaviour(sts, raster_product)
}
\arguments{
\item{sts}{Storms object that contains the storm we are interested in}

\item{raster_product}{raster which spatializes the product we are interested
in. The name of the layer must be nameOfTheStorm_product (which is the case)
if the product is generated with the function \code{stormBehaviour}}
}
\description{
Plot a selected product (MSW, PDI ...) alongside with the track of a selected
storm contained in object \code{sts}.
}
