% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotLabels}
\alias{plotLabels}
\title{Add ISO Times and name labels on the track of a Storm on a map that should be
previsouly plotted}
\usage{
plotLabels(st, by, pos)
}
\arguments{
\item{st}{Storm object}

\item{by}{numeric. Increment of the sequence for the labels to plot}

\item{pos}{numeric. Must be between 1 and 4 and correspond to the position of
labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right)}
}
\description{
Add ISO Times and name labels on the track of a Storm on a map that should be
previsouly plotted
}
