% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTEW.R
\name{plotTemporalTEW}
\alias{plotTemporalTEW}
\title{Plotting TEW time series at given point locations}
\usage{
plotTemporalTEW(data, storm)
}
\arguments{
\item{data}{time series generated with \code{temporalBehaviour} with \code{product=TS}
input first and then \code{computeTEW}}

\item{storm}{list of characters. Storm names. The storm must be available in
\code{data} input. It can also be a vector of integer corresponding to the
indices of storms stored in \code{data}input.}
}
\value{
null
}
\description{
Plotting TEW time series at given point locations
}
