% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotTrack}
\alias{plotTrack}
\title{Plot the track of a Storm on a map that should be previsouly plotted}
\usage{
plotTrack(st, all_basin)
}
\arguments{
\item{st}{Storm object}

\item{all_basin}{logical. Whether or not to plot the track onto the whole basin.
Default value is set to \code{FALSE}. Otherwise, the plot focuses on the extent of
\code{spatial.loi.buffer} of the Storms object in which \code{storm} belongs}
}
\description{
Plot the track of a Storm on a map that should be previsouly plotted
}
