% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defStormsList.R
\name{removeStorms}
\alias{removeStorms}
\title{Removing specific storms from a \code{stormsList} object}
\usage{
removeStorms(sts, names, seasons)
}
\arguments{
\item{sts}{\code{stormsList} object}

\item{names}{character vector. Names of storms to remove (in capital letters)}

\item{seasons}{numeric vector. Seasons of storms to remove (same length as \code{names})}
}
\value{
\code{stormsList} object without the specified storms
}
\description{
The \code{removeStorms()} function removes specified storms from a \code{stormsList}
object based on their names and seasons.
}
\examples{
\dontrun{
sds <- defStormsDataset(...)
sts <- defStormsList(sds, loi="Vanuatu", seasons=c(1990, 2000))

# Remove specific storms
sts <- removeStorms(sts, names=c("UNNAMED", "BARRY"), seasons=c(1990, 1989))
}
}
