% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defStormsList.R
\name{renameStorms}
\alias{renameStorms}
\title{Renaming storms to avoid duplicated names in a \code{stormsList} object}
\usage{
renameStorms(sts)
}
\arguments{
\item{sts}{\code{stormsList} object}
}
\value{
\code{stormsList} object with no duplicated names of storms
}
\description{
The \code{renameStorms()} function renames all storms from a \code{stormsList}
that have the same name. It is very common for "UNNAMED" storms for example.
These storms will be renamed "UNNAMED-YEAR-1", "UNNAMED-YEAR-2", ...
}
\examples{
\dontrun{
#Creating a stormsDataset
sds <- defStormsDataset(...)
sts <- defStormsList(...)
getNames(sts)
## "UNNAMED"   "UNNAMED"   "UNNAMED"   "ALLEN"     "CHARLEY"   "DANIELLE"  "JEANNE"    "UNNAMED" 
## "UNNAMED"   "UNNAMED"   "DENNIS"    ...

sts <- renameStorms(sts)
getNames(sts)
## "UNNAMED-1980-1"   "UNNAMED-1980-2"   "UNNAMED-1980-3"   "ALLEN-1980"     "CHARLEY-1980"  ...
## "UNNAMED-1980-6"   "UNNAMED-1980-7"   "DENNIS-1980"    ...
}
}
