% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{show,Storm-method}
\alias{show,Storm-method}
\alias{show,StormsList-method}
\title{Show a Storm/StormsList}
\usage{
\S4method{show}{Storm}(object)

\S4method{show}{StormsList}(object)
}
\arguments{
\item{object}{Storm/StormList object}
}
\description{
Display the Storm/StormsList object
}
\examples{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
## Display information about Niran in sts 
getStorm(sts, name = "NIRAN")

## Display information about sts 
sts
}
