% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{spatialBehaviour}
\alias{spatialBehaviour}
\title{Compute indicators of storm behaviour}
\usage{
spatialBehaviour(
  sts,
  product = "MSW",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Chen",
  empirical_rmw = FALSE,
  space_res = "2.5min",
  temp_res = 1,
  verbose = 2
)
}
\arguments{
\item{sts}{\code{StormsList}}

\item{product}{character. Product to compute among:
\itemize{
\item \code{"MSW"}: Maximum Sustained Wind
\item \code{"PDI"}: Power Dissipation Index
\item \code{"Exposure"}: hour exposition for wind greater than \code{wind_threshold}
\item \code{"Profiles"}, 2D wind speed structures of wind speed with wind
direction for each observation
}
Default value is set to \code{"MSW"}}

\item{wind_threshold}{numeric vector. Minimal wind threshold(s) \eqn{m.s^{-1}} to
compute \code{"Exposure"} product. Ignored if \code{"Exposure"} is not part of the
products to compute. Default value is set to Saffir Simpson Hurricane Scale
thresholds}

\item{method}{character. Cyclonic model used to compute \code{product}. Must be
either:
\itemize{
\item \code{"Willoughby"}: model based on fits performed on cyclonic observations
\item \code{"Holland"}: model based on both basic cyclonic Physics and parameters
fitting according to cyclonic observations
\item \code{"Boose"}: asymmetric model based on Holland model
}
Default value is set to \code{"Willoughby"} (See \code{Details})}

\item{asymmetry}{character. Indicates which version of asymmetry to use in
the computations (see \code{Details}). Must be either:
\itemize{
\item \code{"Miyazaki"}: based on the formula derived in Miyazaki et al. (1962)
\item \code{"Chen"}:  based on the formula derived in Chen (1994)
\item \code{"None"}: no asymmetry is added
}
Default value is set to \code{"Chen"}. Ignored if \code{method == "Boose"}}

\item{empirical_rmw}{logical. Whether to compute the radius of maximum wind
empirically or using the radius of maximum wind from the observations.
Default value is set to \code{FALSE}. If \code{TRUE}, a formula extracted from
Willoughby et al. 2006 is used to compute rmw}

\item{space_res}{character. Space resolution for the raster(s) to compute.
Either \code{"30sec"}, \code{"2.5min"}, \code{"5min"} or \code{"10min"}. Default value is set
to \code{"2.5min"}}

\item{temp_res}{numeric. Period of time used to interpolate data. Allowed
values are \code{1} (60min), \code{0.75} (45min), \code{0.5} (30min), and \code{0.25} (15min).
Default value is set to \code{1}}

\item{verbose}{numeric. Whether or not the function should display
informations about the process and/or outputs. Allowed values are:
\itemize{
\item \code{0}: Nothing is displayed
\item \code{1}: Informations about the process are displayed
\item \code{2}: Outputs are also displayed
}
Default value is set to \code{2}}
}
\value{
SpatRaster stack which provides the desired product computed,
projected in WGS84 and spanning over the extented LOI of the \code{StormsList} .
Number of layers depends on the number of \code{Storm} available in \code{sts} and
also \code{product} and \code{temp_res} inputs:
\itemize{
\item \code{"MSW"} produces one layer per \code{Storm}. Name of layer is "STORMNAME_MSW"
\item \code{"PDI"} produces one layer per \code{Storm}. Name of layer is "STORMNAME_PDI"
\item \code{"Exposure"} produces one layer for each wind values available
in \code{wind_threshold} and for each \code{Storm}. Name of layers are
"STORMNAME_Exposure_threshold1", "STORMNAME_Exposure_threshold2"...
\item \code{"Profiles"} produces two layers for each observations
(real and interpolated) and each  \code{Storm}. Name of layers are
"STORMNAME_Speed_observation", "STORMNAME_Direction_observation"
}
}
\description{
This function computes/rasterizes products for each \code{Storm} included in a
\code{StormsList}, among Maximum Sustained Wind, Power Dissipation Index, Category
exposure and 2D wind speed structures/direction of wind speed for every
observations
}
\details{
The \code{temp_res} input will perform a linear interpolation
of observations to further compute each 2D wind speed structure at each
interpolated observations. For example, if \code{temp_res == 1}, it will
generate observations every 1hour between the available observations. Doing
so, 2D wind speed structure is computed and stacked for each observations
(available and interpolated) to compute the desired product(s) afterwards.
If \code{product == "Profiles"}, nothing else is performed. Otherwise and
depending on the product(s) to compute, calculations are carried out on the
stack and \code{terra::focal} functions are applied to the final raster(s) to
smooth the result(s).

The Willoughby et al. (2006) model is an empirical model fitted to aircraft
observations. The model considers two regions: inside the eye and at external
radii, for which the wind formulations use different exponents to better match
observations. In this model, the wind speed increases as a power function of the
radius inside the eye and decays exponentially outside the eye after a smooth
polynomial transition across the eyewall (see also Willoughby 1995, Willoughby
et al. 2004).

\eqn{\left\{\begin{aligned}
   v_r &= v_m \times \left(\frac{r}{R_m}\right)^{n} \quad if \quad r < R_m \\
   v_r &= v_m \times \left((1-A) \times e^{-\frac{|r-R_m|}{X1}} + A \times e^{-\frac{|r-R_m|}{X2}}\right) \quad if \quad r \geq R_m \\
   \end{aligned}
   \right.
}

where \eqn{v_r} is the radial wind speed (in \eqn{m.s^{-1}}), \eqn{r} is the
distance to the eye of the storm (in \eqn{km}), \eqn{v_m} is the maximum
sustained wind speed (in \eqn{m.s^{-1}}), \eqn{R_m} is the radius of maximum
sustained wind speed (in \eqn{km}), \eqn{X1 = 287.6 - 1.942 \times v_m +
7.799 \times \ln(R_m) + 1.819 \times |\phi|}, \eqn{X2 = 25}, \eqn{n = 2.1340
+ 0.0077 \times v_m - 0.4522 \times \ln(R_m) - 0.0038 \times |\phi|}, \eqn{A
= 0.5913 + 0.0029 \times v_m - 0.1361 \times \ln(R_m) - 0.0042 \times |\phi|}
and \eqn{A\ge0}, \eqn{\phi} is the latitude of the center of the storm.

The Holland (1980) model, widely used in the literature, is based on the
gradient wind balance in mature tropical cyclones. The wind speed distribution
is computed from the circular air pressure field, which can be derived from
the central and environmental pressure and the radius of maximum winds.

\eqn{v_r = \sqrt{\frac{b}{\rho} \times \left(\frac{R_m}{r}\right)^b \times (p_{oci} - p_c) \times e^{-\left(\frac{R_m}{r}\right)^b} + \left(\frac{r \times f}{2}\right)^2} - \left(\frac{r \times f}{2}\right)}

where \eqn{v_r} is the radial wind speed (in \eqn{m.s^{-1}}), \eqn{r} is the
distance to the eye of the storm (in \eqn{km}), \eqn{R_m} is the radius of
maximum sustained wind speed (in \eqn{km}), \eqn{p_c} is the pressure at the
centre of the storm (in \eqn{Pa}), \eqn{p_{oci}} is the pressure at outermost
closed isobar of the storm (in \eqn{Pa}), \eqn{\rho} is the air density set to
\eqn{1.15 kg.m^{-3}}, \eqn{f = 2 \times 7.29 \times 10^{-5} \sin(\phi)} is the
Coriolis force (in \eqn{N.kg^{-1}}, with \eqn{\phi} being the latitude),
\eqn{b = \frac{\rho \times e \times v_m^2}{p_{oci} - p_c}} is the shape
parameter, with \eqn{e} being the base of natural logarithms ~2.718282 and
\eqn{v_m} the maximum sustained wind speed (in \eqn{m.s^{-1}})

The Boose et al. (2004) model, or “HURRECON” model, is a modification of the
Holland (1980) model (see also Boose et al., 2001). In addition to adding
asymmetry, this model treats of water and land differently, using different
surface friction coefficient for each.

\eqn{v_r = F\left(v_m - S \times (1 - \sin(T)) \times \frac{v_h}{2} \right) \times \sqrt{\left(\frac{R_m}{r}\right)^b \times e^{1 - \left(\frac{R_m}{r}\right)^b}}}

where \eqn{v_r} is the radial wind speed (in \eqn{m.s^{-1}}),
\eqn{v_h} is the storm velocity (in \eqn{m.s^{-1}}),
\eqn{r} is the distance to the eye of the storm (in \eqn{km}),
\eqn{v_m} is the maximum sustained wind speed (in \eqn{m.s^{-1}}),
\eqn{R_m} is the radius of maximum sustained wind speed (in \eqn{km}),
\eqn{p_c} is the pressure at the centre of the storm (\eqn{pressure} in \eqn{Pa}),
\eqn{p_{oci}} is the pressure at outermost closed isobar of the storm (in \eqn{Pa}),
\eqn{\rho = 1.15} is the air density (in \eqn{kg.m^{-3}}),
\eqn{b = \frac{\rho \times e \times v_m^2}{p_{oci} - p_c}} is the shape parameter,
\eqn{F} is a scaling parameter for friction (\eqn{1.0} in water, \eqn{0.8} in land),
\eqn{S} is a scaling parameter for asymmetry (\eqn{1.0}),
\eqn{T} oriented angle (clockwise/counter clockwise in Northern/Southern Hemisphere) between forward trajectory of the storm and a radial line from the eye of the storm to point \eqn{r},
\eqn{S} Asymmetry coefficient (usually set to 1)
}
\examples{
\dontrun{
#Compute MSW product for Pam 2015 in Vanuatu using default settings
msw.pam <- spatialBehaviour(pam)

#Compute PDI product for Storms in sts_nc using Holland model without asymmetry
pdi.nc <- spatialBehaviour(sts_nc, method = "Holland", product = "PDI", asymmetry = "None")

#Compute Exposure for Pam 2015 in Vanuatu using default settings
exp.pam <- spatialBehaviour(pam, product = "Exposure")

#Compute profiles for Storms in sts_nc using default settings
prof.nc <- spatialBehaviour(sts_nc, product = "Profiles")
}

}
