% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sshs}
\alias{sshs}
\title{Saffir Simpson Hurricane Scale}
\format{
\subsection{\code{sshs}}{

a numeric vector of length 7
}
}
\usage{
sshs
}
\description{
Saffir Simpson Hurricane Scale
}
\keyword{datasets}
