% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sshsPalette}
\alias{sshsPalette}
\title{Saffir Simpson Hurricane Scale color Palette}
\format{
\subsection{\code{sshsPalette}}{

a character vector of length 7
}
}
\usage{
sshsPalette
}
\description{
Saffir Simpson Hurricane Scale color Palette
}
\keyword{datasets}
