% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{stormBehaviour_pt}
\alias{stormBehaviour_pt}
\title{Compute indicators of storm behaviour}
\usage{
stormBehaviour_pt(
  sts,
  points,
  product = "TS",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Boose01",
  empirical_rmw = FALSE,
  time_res = 1
)
}
\arguments{
\item{sts}{Storms object}

\item{points}{data.frame. Contains longitude/latitude coordinates within column
names "lon" and "lat", on which to compute the desired product}

\item{product}{character. Product to compute that is either
\itemize{
\item "TS", Time Series of wind speed
\item "PDI", Power Dissipation Index
\item "Exposure", hour exposition for wind upper than wind_threshold input
}
Default value is set to "TS"}

\item{wind_threshold}{numeric vector. Minimal wind threshold(s) (m/s) to compute
Exposure product. Ignored if Exposure is not part of the product to compute.
Default value is set Saffir Simpson Hurricane Scale thresholds}

\item{method}{character. Cyclonic model used to compute product, that is either
\itemize{
\item "Willoughby"
\item "Holland"
}
Default value is set to "Willoughby"}

\item{asymmetry}{character. Indicates which version of asymmetry to use in
the computations, that is either
\itemize{
\item "None", no asymmetry
\item "Boose01", (See Details)
}
Default value is set to "Boose01".}

\item{empirical_rmw}{logical. Whether to compute the Radius of Maximum Wind
empirically, according to getRmw function (See Details) or using the Radius of Maximum Wind
from the observations. Default value is set to FALSE}

\item{time_res}{numeric. Period of time used to interpolated data.
Allowed values are 1 (60min), 0.75 (45min), 0.5 (30min), and 0.25 (15min).
Default value is set to 1}
}
\value{
Computed product for each points are returned through data.frames contained in a named list.
Each slot (named after the storm) is a data.frame that has the following dimensions:
\itemize{
\item If product == "TS": a data frame whose number of rows corresponds to the number
of interpolated obervations. The columns provides respectively the wind speed value (m/s),
the wind direction (in degree), the indice of observation and the ISO time of observation.
\item If product == "PDI": a data.frame with one row and one column that
contains PDI value for each point in points.
\item If product == "Exposure": a data.frame with one row for each wind threshold
and one column that contains Exposure value for each point in points.
}
}
\description{
This function is the pointwise version of stormBehaviour_sp. Available product are
Time Series of wind speed (TS), Power Dissipation Index (PDI) and Exposure
}
\examples{
\dontrun{
pts <- data.frame(lon = c(166.5, 163), lat = c(-22, -19))

#Compute time series of wind speed for ERICA and NIRAN on points provided in pts using default settings
ts_nc <- stormBehaviour_pt(sts_nc, points = pts)

#Compute PDI for ERICA and NIRAN on points provided in pts using default settings
pdiPt_nc <- stormBehaviour_pt(sts_nc, points = pts, product = "PDI")

#Compute Exposure for ERICA and NIRAN on points provided in df using default settings
expPt_nc <- stormBehaviour_pt(sts_nc, points = pts, product = "Exposure", wind_threshold = c(20,30))
}

}
