% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormDatabase.R
\docType{class}
\name{StormDatabase-class}
\alias{StormDatabase-class}
\alias{StormDataBase}
\title{stormDatabase object}
\description{
Choose the database to use within the package's functions
}
\section{Slots}{

\describe{
\item{\code{url}}{character. Relative path to download the database}

\item{\code{name}}{character. Name of the file of the database stored locally}

\item{\code{format}}{character. Either ".nc" or ".csv"}

\item{\code{fields}}{...}
}}

