% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{Storms-class}
\alias{Storms-class}
\alias{Storms}
\title{Gather all the informations to model a set of Storms}
\value{
A S4 object gathering all the above informations
}
\description{
Gather all the informations to model a set of Storms
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of Storm object}

\item{\code{time.period}}{numerics. Range of the cyclonic seasons of Storms available
in \code{data}}

\item{\code{names}}{characters. Names of Storms available in \code{data}}

\item{\code{sshs}}{numerics. Maximum category in the Saffir Simpson Hurricane Scale
of all Storms available in \code{data}}

\item{\code{nb.storms}}{nuleric. Total Number of Storms available in \code{data}}

\item{\code{basin}}{character. Basin in which the Storms have occured}

\item{\code{loi.basin}}{logical. Whether the loi represents the whole basin or not}

\item{\code{spatial.loi}}{sf object. Represents the location of interest
Projection is EPSG:4326}

\item{\code{buffer}}{numeric. Buffer used to extent \code{spatial.loi} (in km)}

\item{\code{spatial.loi.buffer}}{buffer extension of \code{spatial.loi}}
}}

