% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{Storms}
\alias{Storms}
\title{Creating a \code{StormsList} object}
\usage{
Storms(
  sds,
  loi,
  seasons = c(sds@seasons["min"], sds@seasons["max"]),
  names = NULL,
  max_dist = 300,
  remove_TD = TRUE,
  verbose = 2
)
}
\arguments{
\item{sds}{\code{StormsDataset} object.}

\item{loi}{Location of interest. Can be defined using,
\itemize{
\item character, a country name (e.g., "Vanuatu"),
\item numeric vector, a point coordinate (lon, lat in decimal degrees, e.g., c(169.5, -19.2)), or
\item sp (SpatialPolygon) or a sf (simple features) object (e.g., created from a shapefile).
}}

\item{seasons}{numeric vector. Season(s) of occurrence of the storms. One (e.g., 2020)
or two years (e.g., c(2020,2022)) to extract storms occurring during one specific season or
during several consecutive seasons, respectively. By default all storms occurring since 1980 are extracted.}

\item{names}{character vector. Names of specific storms (in capital letters) to extract.}

\item{max_dist}{numeric. Maximum distance between the location of interest and the storm for which track data are extracted.
Default \code{max_dist} is set to 300 km.}

\item{remove_TD}{logical. Whether (TRUE) or not (FALSE) removing tropical depressions (msw < 18
m/s). Default value is set to \code{TRUE}.}

\item{verbose}{numeric. Type of information the function displays. Can be:
\itemize{
\item \code{2}, information about both the processes and the outputs are displayed (default value),
\item \code{1}, only information about the processes are displayed, or
\item \code{0}, nothing is displayed.
}}
}
\value{
The \code{Storms()} function returns a \code{StormsList} object containing track data for all storms
meeting the specified criteria (e.g., name, season, location).
}
\description{
The \code{Storms()} function extracts storm track data from a \code{StormsDataset}
and creates a \code{StormsList} object.
}
\details{
The available countries for the \code{loi} are those provided in the
\code{rwolrdxtra} package. This package provide high resolution vector country
boundaries derived from Natural Earth data. More informations on the Natural Earth data
here: http://www.naturalearthdata.com/downloads/10m-cultural-vectors/.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting data using country names
vanuatu.st <- Storms(sds = sds, loi = "Vanuatu")

#Getting data using a specific point location
pt <- c(169, -19)
pam.pt <- Storms(sds = sds, loi = pt, names = "PAM")

#Getting data using country and storm names
niran.nc <- Storms(sds = sds, loi = "New Caledonia", names = c("NIRAN"))

#Getting data using a user defined spatial polygon
poly <- cbind(c(135, 290, 290, 135, 135),c(-60, -60, 0, 0, -60))
sp <- sf::st_polygon(list(poly))
sp <- sf::st_sfc(sp, crs = 4326)
sp <- sf::st_as_sf(sp)
sts_sp <- Storms(sds = sds, loi = sp)
}


}
