% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsDataset.R
\docType{class}
\name{StormsDataset-class}
\alias{StormsDataset-class}
\alias{StormsDataset}
\title{StormsDataset object}
\description{
Choose the database to use within the package's functions
}
\details{
The fields input must provide at least 6 mandatory fields (and at most 11) in
order to benefit from all the functionalities of this package:
\itemize{
\item A field called "name": which dimension contains the names of storms
in the netcdf database
\item A field called "seasons": which dimension contains the cyclonic
seasons of storms in the netcdf database
\item A field called "isoTime": which dimension contains the ISO times of
each (3 or 6 hourly) observations for all storms in the database
\item A field called "lon": which dimension contains the longitude
coordinates (Eastern degree) of each observations for all storms in
the netcdf database
\item A field called "lat": which dimension contains the latitude
coordinates (Nothern degree) of each observations for all storms in
the netcdf database
\item A field called "msw": which dimension contains the maximum sustained
wind speed (knt) of each observations for all storms in the netcdf
database
}
The following fields are optional but highly recommanded:
\itemize{
\item A field called "basin": which dimension contains the basin location of
storms in the netcdf database. Used to filter the storms in the netcdf
database
\item A field called "rmw": which dimension contains the radius of maximum
wind speed (nm) of each observations for all storms in the netcdf
database (See spatialBehaviour, temporalBehaviour)
\item A field called "sshs": which dimension contains the Saffir Simpson
Hurricane Scale index of each observations for all storms in the
netcdf database
}
Finally these following fields are optional but mandatory to perform Holland
model (See spatialBehaviour, temporalBehaviour)
\itemize{
\item A field called "pressure": which dimension contains the pressure (mb)
in the eye for of each observations for all storms in the netcdf
database
\item A field called "poci": which dimension contains the Pressure at the
Outermost Closed Isobar (mb) for of each observations for all storms
in the nectdf database
}

Default value is set according to the most relevant dimensions of IBTraCS
databases: fields = fields = c("basin" = "basin", "names" = "name",
"seasons" = "season", "isoTime" = "iso_time", "lon" = "usa_lon",
"lat" = "usa_lat", "msw" = "usa_wind", "rmw" = "usa_rmw",
"pressure" = "usa_pres", "poci" = "usa_poci", "sshs" = "usa_sshs")
}
\section{Slots}{

\describe{
\item{\code{filename}}{character. Name of the database to load. Must be a netcdf file}

\item{\code{fields}}{named character vector. Dictionary that provides all the name of
dimensions to extract from the netcdf database (See Details)}

\item{\code{basin}}{character. Basin name to filter the database within its
boundaries. Default value is set to NULL. It must be either
\itemize{
\item "NA": North Atlantic
\item "SA": South Atlantic
\item "EP": Eastern North Pacific
\item "WP": Western North Pacific
\item "SP": South Pacific
\item "SI": South India
\item "NI": North India
}}

\item{\code{database}}{list of 6 to 10 slots depending on the fields input. Each slot
is either a 1D array of dimension (number of storms), for "names" and
"seasons" fields, or a 2D array of dimension
(Maximum number of observations:number of storms), for the remaining fields
which are "isoTime", "lon", "lat", "msw", "rmw", "pressure", "poci", "sshs"}
}}

