% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sts_nc}
\alias{sts_nc}
\title{ERICA and NIRAN tropical cyclones}
\format{
\subsection{\code{sts_nc}}{

a Storms object generated with getStorms
}
}
\usage{
sts_nc
}
\description{
Gathers informations about both the category 4 and 5 TC ERICA (2003) and
NIRAN (2021) over New Caledonia with a buffer of 300km
}
\keyword{datasets}
