% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sts_wp}
\alias{sts_wp}
\title{Tropical cyclones over the West Pacific Basin}
\format{
\subsection{\code{sts_wp}}{

a Storms object generated with getStorms
}
}
\usage{
sts_wp
}
\description{
Gathers informations about all TCs over category TD in the sshs that occured
in the WP basin between 2010 and 2020
}
\keyword{datasets}
