% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defStormsList.R
\name{subsetStorms}
\alias{subsetStorms}
\title{Keeping only specific storms from a \code{stormsList} object}
\usage{
subsetStorms(sts, names, seasons)
}
\arguments{
\item{sts}{\code{stormsList} object}

\item{names}{character vector. Names of storms to keep (in capital letters)}

\item{seasons}{numeric vector. Seasons of storms to keep (same length as \code{names})}
}
\value{
\code{stormsList} object containing only the specified storms
}
\description{
The \code{subsetStorms()} function keeps only specified storms from a \code{stormsList}
object based on their names and seasons, removing all others.
}
\examples{
\dontrun{
sds <- defStormsDataset(...)
sts <- defStormsList(sds, loi="Vanuatu", seasons=c(1990, 2000))

# Keep only specific storms
sts <- subsetStorms(sts, names=c("PAM", "NIRAN"), seasons=c(2015, 2021))
}
}
