% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{temporalBehaviour}
\alias{temporalBehaviour}
\title{Compute indicators of storm behaviour}
\usage{
temporalBehaviour(
  sts,
  points,
  product = "TS",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Chen",
  empirical_rmw = FALSE,
  temp_res = 1,
  verbose = 1
)
}
\arguments{
\item{sts}{\code{StormsList} object}

\item{points}{data.frame. Contains longitude/latitude coordinates within
column names \code{x} and \code{y}, on which to compute the desired product}

\item{product}{character. Product to compute. Must be either:
\itemize{
\item \code{"TS"}: Time Series of wind speed
\item \code{"PDI"}: Power Dissipation Index
\item \code{"Exposure"}: hour exposition for wind greater than \code{wind_threshold}
input
}
Default value is set to \code{"TS"}}

\item{wind_threshold}{numeric vector. Minimal wind threshold(s) (m/s) to
compute \code{"Exposure"} product. Ignored if \code{"Exposure"} is not part of the
products to compute. Default value is set to Saffir Simpson Hurricane Scale
thresholds}

\item{method}{character. Cyclonic model used to compute product. Must be
either:
\itemize{
\item \code{"Willoughby"}: model based on fits performed on cyclonic observations
\item \code{"Holland"}: model based on both basic cyclonic Physics and parameters
fitting according to cyclonic observations
\item \code{"Boose"}: asymmetric model based on Holland model
}
Default value is set to \code{"Willoughby"} (See \code{Details})}

\item{asymmetry}{character. Indicates which version of asymmetry to use in
the computations (see \code{Details}). Must be either:
\itemize{
\item \code{"Miyazaki"}: based on the formula derived in Miyazaki et al. (1962)
\item \code{"Chen"}:  based on the formula derived in Chen (1994)
\item \code{"None"}: no asymmetry is added
}
Default value is set to \code{"Chen"}. Ignored if \code{method == "Boose"}}

\item{empirical_rmw}{logical. Whether to compute the radius of maximum wind
empirically or using the radius of maximum wind from the observations.
Default value is set to \code{FALSE}. If \code{TRUE}, a formula extracted from
Willoughby et al. 2006 is used to compute rmw}

\item{temp_res}{numeric. Period of time used to interpolate data. Allowed
values are \code{1} (60min), \code{0.75} (45min), \code{0.5} (30min), and \code{0.25} (15min).
Default value is set to \code{1}}

\item{verbose}{numeric. Whether or not the function should display
informations about the process and/or outputs and additional notes. Allowed
values are:
\itemize{
\item \code{0}: Nothing is displayed
\item \code{1}: Informations about the process are displayed
\item \code{2}: Outputs are also displayed
\item \code{3}: Additional notes are also displayed
}
Default value is set to \code{2}}
}
\value{
Computed product for each points are returned through lists of
data.frame (one per points coordinates) contained in a named list. Then,
each slot, named after the storm, is made of list(s) of data.frame that has
the following dimensions:
\itemize{
\item If \code{product == "TS"}: a data frame whose number of rows corresponds to
the number of interpolated observations. The columns provides
respectively the wind speed values (m/s), the wind directions
(in degree), the indices and the ISO time of observations
\item If \code{product == "PDI"}: a data.frame with one row and one column that
contains PDI value for each point in points.
\item If \code{product == "Exposure"}: a data.frame with one row for each wind
threshold and one column that contains Exposure value for each point
in points.
}
}
\description{
This function is the pointwise version of \code{spatialBehaviour}. Available
products are Time Series of wind speed (TS), Power Dissipation Index (PDI)
and Exposure
}
\details{
The \code{temp_res} input will perform a linear interpolation of
observations to further compute each 2D wind speed structure at each
interpolated observations. For example, if \code{temp_res == 1}, it will
generate observations every 1hour between the available observations.
}
\examples{
\dontrun{
pts <- data.frame(lon = c(166.5, 163), lat = c(-22, -19))

#Compute time series of wind speed for ERICA and NIRAN on points
#provided in pts using default settings
ts_nc <- temporalBehaviour(sts_nc, points = pts)

#Compute PDI for ERICA and NIRAN on points provided in pts using default settings
pdiPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "PDI")

#Compute Exposure for ERICA and NIRAN on points provided in df using default settings
expPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "Exposure", wind_threshold = c(20,30))
}

}
