% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{Willoughby_profil}
\alias{Willoughby_profil}
\title{Compute wind values according to the Willoughby model}
\usage{
Willoughby_profil(w_max, lat, r, rmw = NULL)
}
\arguments{
\item{w_max}{Maximum wind speed generated}

\item{lat}{Latitude of the eye of the storm}

\item{r}{Distance to the center of the storm where the value must be computed}

\item{rmw}{Radius of maximum wind speed. Default value is set to NULL}
}
\value{
wind value according to the Willoughby model at distance \code{r} to the
center of the storm located in latitude \code{lat}
}
\description{
Compute wind values according to the Willoughby model
}
