% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{Willoughby_profile}
\alias{Willoughby_profile}
\title{Compute radial wind speed according to Willoughby model}
\usage{
Willoughby_profile(r, rmw, msw, lat)
}
\arguments{
\item{r}{numeric. Distance to the eye of the storm (km) where the value must be computed}

\item{rmw}{numeric. Radius of Maximum Wind (km)}

\item{msw}{numeric. Maximum Sustained Wind (m/s)}

\item{lat}{numeric. Should be between -60 and 60. Latitude of the eye of the storm}
}
\value{
radial wind speed value (m/s) according to Willoughby model at distance \code{r} to the
eye of the storm located in latitude \code{lat}
}
\description{
Compute radial wind speed according to Willoughby model
}
