% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRast.R
\name{writeRast}
\alias{writeRast}
\title{Save a raster in the desired format}
\usage{
writeRast(rast, format = ".tiff", filename = NULL, path = "./")
}
\arguments{
\item{rast}{SpatRast object.}

\item{format}{character. Format of the file to export \code{rast}. Either \code{.tiff}
or \code{.nc}. Default value is set to \code{.tiff}}

\item{filename}{character. Name of the file. Default value is set to \code{NULL},
in this case it will be set to \code{names(rast)}}

\item{path}{character. Relative path where the file should be written}
}
\description{
This function writes a SpatRast raster in the given format
}
