% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{Boose01_profile}
\alias{Boose01_profile}
\title{Compute wind values according to the Boose 01 model}
\usage{
Boose01_profile(r, t, rmw, landfall, w_max, Vh, b = 1.3)
}
\arguments{
\item{r}{Distance to the center of the storm where the value must be computed}

\item{t}{Clockwise angle between forward motion of hurricane and radial line to P}

\item{rmw}{Radius of maximum wind speed. Default value is set to NULL}

\item{landfall}{pres Pressure at the center of the storm}

\item{w_max}{Maximum wind speed generated}

\item{Vh}{forward speed of storm}

\item{b}{Scaling factor}
}
\value{
wind value according to the Boose 01 at distance \code{r} to the
center of the storm
}
\description{
Compute wind values according to the Boose 01 model
}
