% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cyclone_energy.R
\name{calculate_cyclone_energy}
\alias{calculate_cyclone_energy}
\title{Calculate the Accumulated Cyclone Energy of a Storm}
\usage{
calculate_cyclone_energy(df)
}
\arguments{
\item{df}{A dataframe containing the storm data for a single storm.}
}
\value{
A numeric value representing the accumulated cyclone energy of the storm.
}
\description{
This function computes the accumulated cyclone energy of a given storm.
}
