% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTEW.R
\name{computeTEW}
\alias{computeTEW}
\title{Compute the topographic exposure to wind}
\usage{
computeTEW(profiles, ...)
}
\description{
The \code{computeTEW()} function allows computing topographic exposure to wind (TEW)
for either:
\enumerate{
\item each cell of a regular grid (i.e., a raster)
or
\item a data.frame with the coordinates of the points of interest,
for a given tropical cyclone or a set of tropical cyclones.
}
}
