% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsDataset.R
\name{defDatabase}
\alias{defDatabase}
\title{Initialize a StormsDataset}
\usage{
defDatabase(
  filename,
  fields = c(names = "name", seasons = "season", isoTime = "iso_time", lon = "usa_lon",
    lat = "usa_lat", msw = "usa_wind", basin = "basin", sshs = "usa_sshs", rmw =
    "usa_rmw", pressure = "usa_pres", poci = "usa_poci"),
  basin = NULL,
  seasons = c(1980, as.numeric(format(Sys.time(), "\%Y"))),
  unit_conversion = c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci =
    "mb_to_pa"),
  verbose = TRUE
)
}
\arguments{
\item{filename}{character. Name of the database to load. Must be a netcdf
file}

\item{fields}{named character vector. Dictionary that provides all the name
of dimension to extract from the netcdf database (See \code{StormsDataSet}
class)}

\item{basin}{character. Basin name to filter the database within its
boundaries. Default value is set to \code{NULL}. Otherwise, it must be either
\itemize{
\item \code{"NA"}: North Atlantic
\item \code{"SA"}: South Atlantic
\item \code{"EP"}: Eastern North Pacific
\item \code{"WP"}: Western North Pacific
\item \code{"SP"}: South Pacific
\item \code{"SI"}: South India
\item \code{"NI"}: North India
}}

\item{seasons}{numeric vector. Range of calendar years to filter storms. For
cyclones that formed in one year and dissipated in the following year, the
latter should be used. Default value is set to between \code{1980} and the
current year}

\item{unit_conversion}{named character vector. Dictionary that provides the
instructions to convert \code{msw}, \code{rmw}, \code{pressure} and/or \code{poci} fields in
the correct units. Default value is \code{c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci = "mb_to_pa")} according to the IBTrACS
databases. This input is mandatory even if no conversion is needed. In this
case use \code{"None"} in the corresponding fields. Allowed value are: For \code{msw}
fields
\itemize{
\item \code{"knt_to_ms"}: converts knot in meter per second
\item \code{"kmh_to_ms"}: converts kilometer per hour in meter per second
\item "\verb{mph_to_ms"}: converts miles per hour in meter per second
\item \code{"None"}: No conversion needed
}

For \code{rmw} fields
\itemize{
\item \code{"nm_to_ms"}: converts nautical miles in kilometer
\item \code{"None"}: No conversion needed
}

For both \code{pressure} and \code{poci}
\itemize{
\item \code{"b_to_pa"}: converts bar in pascal
\item "\verb{mb_to_pa"}: converts millibar in pascal
\item \code{"atm_to_pa"}: converts atmosphere in pascal
\item \code{"psi_to_pa"}: converts psi in pascal
\item \code{"None"}: No conversion needed
}}

\item{verbose}{logical. Whether or not the function should display
informations about the process}
}
\value{
An object of class \code{StormsDataset}
}
\description{
Initialize a StormsDataset
}
