% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsDataset.R
\name{defDatabase}
\alias{defDatabase}
\title{The \code{defDatabase()} function creates a \code{StormsDataset} object from a NetCDF file}
\usage{
defDatabase(
  filename,
  fields = c(names = "name", seasons = "season", isoTime = "iso_time", lon = "usa_lon",
    lat = "usa_lat", msw = "usa_wind", sshs = "usa_sshs", rmw = "usa_rmw", pressure =
    "usa_pres", poci = "usa_poci"),
  basin = NULL,
  seasons = c(1980, as.numeric(format(Sys.time(), "\%Y"))),
  unit_conversion = c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci =
    "mb_to_pa"),
  verbose = TRUE
)
}
\arguments{
\item{filename}{character. Name of the NetCDF (.nc) file.}

\item{fields}{named character vector. Correspondence between the fields
in the output \code{StormsDataset} object and the variable names in the input NetCDF file.}

\item{basin}{character. Name of the basin for which storm track data are extracted.
By default \code{basin=NULL}, meaning that all stars regardless the basin in which they
originated are extracted. Seven basins can be used to filter the data set:
\itemize{
\item \code{"NA"} for North Atlantic basin.
\item \code{"SA"} for South Atlantic basin.
\item \code{"EP"} for Eastern North Pacific basin.
\item \code{"WP"} for Western North Pacific basin.
\item \code{"SP"} for South Pacific basin.
\item \code{"SI"} for South India basin.
\item \code{"NI"} for North India basin.
}}

\item{seasons}{numeric vector. Seasons of occurrence of the storms (e.g., c(2020,2022)).
By default all storms occurring since 1980 are extracted.}

\item{verbose}{logical. Whether (TRUE) or not (FALSE) the function should display
information about the processes}
}
\value{
The \code{defDatabase()} function returns a \code{StormsDataset} object.
}
\description{
The \code{defDatabase()} function creates a \code{StormsDataset} object from a NetCDF file
}
