% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsDataset.R
\name{defDatabase}
\alias{defDatabase}
\title{Initialize a StormsDataset object}
\usage{
defDatabase(
  filename,
  fields = c(basin = "basin", names = "name", seasons = "season", isoTime = "iso_time",
    lon = "usa_lon", lat = "usa_lat", msw = "usa_wind", sshs = "usa_sshs", rmw =
    "usa_rmw", pressure = "usa_pres", poci = "usa_poci"),
  basin = "SP",
  unit_conversion = c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci =
    "mb_to_pa"),
  verbose = TRUE
)
}
\arguments{
\item{filename}{character. Name of the database to load. Must be a netcdf
file}

\item{fields}{named character vector. Dictionary that provides all the name
of dimension to extract from the netcdf database (See StormsDataSet class)}

\item{basin}{character. Basin name to filter the database within its
boundaries. Default value is set to NULL. It must be either
\itemize{
\item "NA": North Atlantic
\item "SA": South Atlantic
\item "EP": Eastern North Pacific
\item "WP": Western North Pacific
\item "SP": South Pacific
\item "SI": South India
\item "NI": North India
}}

\item{unit_conversion}{...}

\item{verbose}{logical. Whether or not the function should display
informations about the process}
}
\value{
An object of class StormsDataset
}
\description{
Initialize a StormsDataset object
}
