% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsDataset.R
\name{defDatabase}
\alias{defDatabase}
\title{The \code{defDatabase()} function creates a \code{StormsDataset} object from a NetCDF file}
\usage{
defDatabase(
  filename = system.file("extdata", "test_dataset.nc", package = "StormR"),
  fields = c(names = "name", seasons = "season", isoTime = "iso_time", lon = "usa_lon",
    lat = "usa_lat", msw = "usa_wind", sshs = "usa_sshs", rmw = "usa_rmw", pressure =
    "usa_pres", poci = "usa_poci"),
  basin = NULL,
  seasons = c(1980, as.numeric(format(Sys.time(), "\%Y"))),
  unit_conversion = c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci =
    "mb_to_pa"),
  verbose = TRUE
)
}
\arguments{
\item{filename}{character. Name of the NetCDF (.nc) file. Default is the \code{test_dataset.nc}
file located in the \code{inst/extdata} repository of the directory (accessible by
\code{system.file("extdata", "test_dataset.nc", package = "StormR")}). This toy dataset is extracted
from the IBTrACS.SP.v04r00.nc file. Provides all the tropical cyclones that occured around Vanuatu
from 2015 to 2016 and around New Caledonia from 2020 to 2021.}

\item{fields}{named character vector. Correspondence between the fields
in the output \code{StormsDataset} object for internal representation in \code{StormR} package
and the variable names in the input NetCDF file.
By default, they are chosen to meet the mapping when importing a NetCDF file from the IBTrACS database.}

\item{basin}{character. Name of the basin for which storm track data are extracted.
By default \code{basin=NULL}, meaning that all stars regardless the basin in which they
originated are extracted. Seven basins can be used to filter the data set:
\itemize{
\item \code{"NA"} for North Atlantic basin.
\item \code{"SA"} for South Atlantic basin.
\item \code{"EP"} for Eastern North Pacific basin.
\item \code{"WP"} for Western North Pacific basin.
\item \code{"SP"} for South Pacific basin.
\item \code{"SI"} for South India basin.
\item \code{"NI"} for North India basin.
}}

\item{seasons}{numeric vector. Seasons of occurrence of the storms (e.g., c(2020,2022)).
By default all storms occurring since 1980 are extracted.}

\item{unit_conversion}{named character vector. Required unit conversions, \code{msw} has
to be provided in $m.s^{-1}$, \code{rmw} in $km$, \code{pressure} and \code{poci} in $Pa$. By default
\code{unit_conversion=c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci = "mb_to_pa")}
to meet the conversion requirement when importing a NetCDF file from the IBTrACS database.
This argument is mandatory even if no conversion is needed. If no conversion is needed then
use \code{"None"} in the corresponding fields. The following unit conversions are implemented:

For \code{msw},
\itemize{
\item \code{"knt_to_ms"} to convert knot to meter per second (default setting).
\item \code{"kmh_to_ms"} to convert kilometre per hour to meter per second.
\item "\verb{mph_to_ms"} to convert miles per hour to meter per second.
\item \code{"None"}if no conversion is needed.
}

For \code{rmw},
\itemize{
\item \code{"nm_to_ms"}to convert nautical miles to kilometre (default setting).
\item \code{"None"}if no conversion is needed.
}
For \code{pressure} and \code{poci},
\itemize{
\item "\verb{mb_to_pa"} to convert  millibar to Pascal  (default setting).
\item \code{"b_to_pa"} to convert bar to Pascal.
\item \code{"atm_to_pa"} to convert  atmosphere to Pascal.
\item \code{"psi_to_pa"} to convert  psi to Pascal.
\item \code{"None"}if no conversion is needed.
}}

\item{verbose}{logical. Whether (TRUE) or not (FALSE) the function should display
information about the processes}
}
\value{
The \code{defDatabase()} function returns a \code{StormsDataset} object.
}
\description{
The \code{defDatabase()} function creates a \code{StormsDataset} object from a NetCDF file
}
