% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getBuffer}
\alias{getBuffer}
\alias{getBuffer,StormsList-method}
\title{Getting the buffered location of interest}
\usage{
getBuffer(sts)

\S4method{getBuffer}{StormsList}(sts)
}
\arguments{
\item{sts}{\code{StormsList}}
}
\value{
The \code{getBuffer()} returns a \code{sf} object.
}
\description{
The \code{getBuffer()} function allows to get the buffered location of interest from a \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the buffered location of interest from the sts object
buff <- getBuffer(sts)
}
}
