% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getBuffer}
\alias{getBuffer}
\alias{getBuffer,StormsList-method}
\title{Get the extended LOI from a Storms object}
\usage{
getBuffer(sts)

\S4method{getBuffer}{StormsList}(sts)
}
\arguments{
\item{sts}{StormsList object}
}
\value{
sf object. Extended Location Of Interest for the given StormsList
object
}
\description{
Get the Extended Location Of Interest from a StormsList object
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
buff <- getBuffer(sts)
}
}
