% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getBuffer}
\alias{getBuffer}
\alias{getBuffer,Storms-method}
\title{Get the extended location of interest from a Storms object}
\usage{
getBuffer(sts)

\S4method{getBuffer}{Storms}(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
sf object. Extended location of interest for the given Storms object
}
\description{
Get the Extended location of interest of a Storms object
}
