% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getBufferSize}
\alias{getBufferSize}
\alias{getBufferSize,StormsList-method}
\title{Getting the buffer size}
\usage{
getBufferSize(sts)

\S4method{getBufferSize}{StormsList}(sts)
}
\arguments{
\item{sts}{\code{StormsList}}
}
\value{
numeric.
}
\description{
The \code{getBufferSize()} returns the buffer size used to generate the buffered location of interest for a \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the buffer size from the sts object
buffsize <- getBufferSize(sts)
}
}
