% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getBufferSize}
\alias{getBufferSize}
\alias{getBufferSize,Storms-method}
\title{getBufferSize}
\usage{
getBufferSize(sts)

\S4method{getBufferSize}{Storms}(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
numeric. Buffer size (in km) used to generate the extended location
of interest for the given Storms object
}
\description{
Get the buffer size for a Storms object
}
