% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{getColors}
\alias{getColors}
\title{Get SSWS palette colors associated with a wind observation}
\usage{
getColors(ms_wind)
}
\arguments{
\item{ms_wind}{maximum sustained wind observation}
}
\value{
color palette associated with the observation
}
\description{
Get SSWS palette colors associated with a wind observation
}
