% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getInObs}
\alias{getInObs}
\alias{getInObs,Storms-method}
\alias{getInObs,Storm-method}
\title{getInObs}
\usage{
getInObs(s, ...)

\S4method{getInObs}{Storms}(s, name, season = NULL)

\S4method{getInObs}{Storm}(s)
}
\arguments{
\item{s}{Storms or Storm object}

\item{...}{Additional arguments}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the Storms object share the same name. Default value is set
to NULL}
}
\value{
numeric vector. Indices within the extended LOI for the given Storm object
}
\description{
Get the indices of observations within th extended Location Of Interest for
a given storm
}
