% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getInObs}
\alias{getInObs}
\alias{getInObs,StormsList-method}
\alias{getInObs,Storm-method}
\title{Get indices of observations within the extended LOI}
\usage{
getInObs(s, ...)

\S4method{getInObs}{StormsList}(s, name, season = NULL)

\S4method{getInObs}{Storm}(s)
}
\arguments{
\item{s}{Storm/StormsList object}

\item{...}{extra argument depending on Storm/StormsList object}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the StormsList object share the same name. Default value is set
to NULL}
}
\value{
numeric vector. Indices within the extended LOI for the given Storm object
}
\description{
Get the indices of observations within the extended Location Of Interest for
a given storm
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))

## For storm Niran
getInObs(getStorm(sts, name = "NIRAN"))

## Equivalent to 
getInObs(sts, name = "NIRAN")
}
}
