% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getInObs}
\alias{getInObs}
\alias{getInObs,StormsList-method}
\alias{getInObs,Storm-method}
\title{Getting the number of the observations}
\usage{
getInObs(s, ...)

\S4method{getInObs}{StormsList}(s, name, season = NULL)

\S4method{getInObs}{Storm}(s)
}
\arguments{
\item{s}{\code{Storm} or \code{StormsList} object.}

\item{...}{extra argument for \code{StormsList}}

\item{name}{character. Name of the storm in capital letters.}

\item{season}{numeric. Cyclonic season of the \code{Storm}. Required only
if several \code{Storm} in \code{s} object have the same name. Default value is set
to \code{NULL}}
}
\value{
numeric vector. The \code{getInObs()} function returns the number of the observations in a given \code{Storm} or \code{StormsList} object.
}
\description{
The \code{getInObs()} function allows to get the number of the of the observations in a given \code{Storm} or \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the number of the observation for the tropical cyclone Niran in the sts object 
getInObs(getStorm(sts, name = "NIRAN"))
getInObs(sts, name = "NIRAN")
}
}
