% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getLOI}
\alias{getLOI}
\title{Get LOI from a Storms object}
\usage{
getLOI(sts)
}
\arguments{
\item{sts}{StormsList object}
}
\value{
sf object. Location Of Interest for the given StormsList object
}
\description{
Get the Location Of Interest from a StormsList object
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
loi <- getLOI(sts)
}
}
