% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getLOI}
\alias{getLOI}
\alias{getLOI,StormsList-method}
\title{Getting the location of interest}
\usage{
getLOI(sts)

\S4method{getLOI}{StormsList}(sts)
}
\arguments{
\item{sts}{\code{StormsList} object}
}
\value{
sf object. The \code{getLOI()} functions returns the location of interest for the given \code{StormsList}.
}
\description{
The \code{getLOI()} functions allows to get the location of interest from a \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the location of interest for the sts object
loi <- getLOI(sts)
}
}
