% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNames}
\alias{getNames}
\alias{getNames,Storm-method}
\alias{getNames,StormsList-method}
\title{Getting the names of the storms}
\usage{
getNames(s)

\S4method{getNames}{Storm}(s)

\S4method{getNames}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm} or \code{StormsList} object.}
}
\value{
character vector.
}
\description{
The \code{getNames()} function returns the names of the storms in a \code{Storm} or a \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the names of the storms from the sts object
getNames(sts)
}
}
