% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNames}
\alias{getNames}
\alias{getNames,Storm-method}
\alias{getNames,StormsList-method}
\title{Get name(s) of storm(s)}
\usage{
getNames(s)

\S4method{getNames}{Storm}(s)

\S4method{getNames}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm} / \code{StormsList}}
}
\value{
character vector. Names of each storms provided by the given \code{Storm}
/ \code{StormsList}
}
\description{
Get the name(s) of storm(s) available in a \code{Storm} / \code{StormsList}
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")

## For storm Niran
getNames(getStorm(sts, name = "NIRAN"))

## For sts
getNames(sts)
}
}
