% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNames}
\alias{getNames}
\alias{getNames,Storms-method}
\alias{getNames,Storm-method}
\title{Get the name(s) of storm(s) provided in a Storms/Storm object}
\usage{
getNames(s)

\S4method{getNames}{Storms}(s)

\S4method{getNames}{Storm}(s)
}
\arguments{
\item{s}{Storms or Storm object}
}
\value{
character vector. Names of each storms provided by the given Storms
object or name of the storm of the given Storm object
}
\description{
Get the names of storms available in a Storms object or the name of the storm
of a Storm object
}
