% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNames}
\alias{getNames}
\title{Get name(s) of storm(s)}
\usage{
getNames(s)
}
\arguments{
\item{s}{Storm/StormsList object}
}
\value{
character vector. Names of each storms provided by the given Storm/StormsList
object
}
\description{
Get the name(s) of storm(s) available in a Storm/StormsList object
}
\examples{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))

## For storm Niran
getNames(getStorm(sts, name = "NIRAN"))

## For sts
getNames(sts)
}
