% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbObs}
\alias{getNbObs}
\alias{getNbObs,Storms-method}
\title{getNbobs}
\usage{
getNbObs(sts, name, season = NULL)

\S4method{getNbObs}{Storms}(sts, name, season = NULL)
}
\arguments{
\item{sts}{Storms object}

\item{name}{character. Name of the storm to extract}

\item{season}{numeric. Cyclonic season of the storm to extract. Use only if several
storms in the Storms object share the same name. Default value is set de NULL}
}
\value{
numeric. Number of observations available within the given storm
}
\description{
getNbobs
}
