% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbObs}
\alias{getNbObs}
\alias{getNbObs,Storm-method}
\alias{getNbObs,StormsList-method}
\title{Getting the number of observations}
\usage{
getNbObs(s, ...)

\S4method{getNbObs}{Storm}(s)

\S4method{getNbObs}{StormsList}(s, name, season = NULL)
}
\arguments{
\item{s}{\code{Storm} or \code{StormsList} object.}

\item{...}{extra arguments for \code{StormsList}}

\item{name}{character. Name of the storm in capital letters.}

\item{season}{numeric. Cyclonic season of the \code{Storm}. Required only
if several \code{Storm} in the \code{s} have the same name. Default value is set to
\code{NULL}.}
}
\value{
numeric. The getNbObs() function returns the number of observations for a storm in a \code{Storm} or \code{StormsList} object.
}
\description{
The getNbObs() function allows to get the number of observations available for a storm in a \code{Storm} or \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

##Getting the number of observations for the tropical cyclone Niran in the sts object 
getNbObs(getStorm(sts, name = "NIRAN"))
getNbObs(sts, name = "NIRAN")
}
}
