% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbObs}
\alias{getNbObs}
\alias{getNbObs,Storms-method}
\alias{getNbObs,Storm-method}
\title{getNbobs}
\usage{
getNbObs(s, ...)

\S4method{getNbObs}{Storms}(s, name, season = NULL)

\S4method{getNbObs}{Storm}(s)
}
\arguments{
\item{s}{Storms or Storm object}

\item{...}{Additional arguments}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the Storms object share the same name. Default value is set
to NULL}
}
\value{
numeric. Number of observations available for the given storm
}
\description{
Get the number of observations available for a given storm
}
