% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbObs}
\alias{getNbObs}
\alias{getNbObs,Storm-method}
\alias{getNbObs,StormsList-method}
\title{Get the number of observations available}
\usage{
getNbObs(s, ...)

\S4method{getNbObs}{Storm}(s)

\S4method{getNbObs}{StormsList}(s, name, season = NULL)
}
\arguments{
\item{s}{Storm/StormsList object}

\item{...}{extra argument depending on Storm/StormsList object}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the StormsList object share the same name. Default value is set
to NULL}
}
\value{
numeric. Number of observations available for the given storm
}
\description{
Get the number of observations available for a given storm
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))

## For storm Niran
getNbObs(getStorm(sts, name = "NIRAN"))

## Equivalent to 
getNbObs(sts, name = "NIRAN")
}
}
