% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbStorms}
\alias{getNbStorms}
\alias{getNbStorms,stormsList-method}
\title{Getting the number of \code{storm}}
\usage{
getNbStorms(sts)

\S4method{getNbStorms}{stormsList}(sts)
}
\arguments{
\item{sts}{\code{stormsList}}
}
\value{
numeric.
}
\description{
The \code{getNbStorms()} returns the number of \code{storm} objects
in the given \code{stormsList} object.
}
\examples{
\dontrun{
#Creating a stormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- storms(sds=sds, loi = "New Caledonia")

#Getting the number of storms in the sts object
getNbStorms(sts)
}
}
