% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbStorms}
\alias{getNbStorms}
\alias{getNbStorms,StormsList-method}
\title{Getting the number of \code{Storm}}
\usage{
getNbStorms(sts)

\S4method{getNbStorms}{StormsList}(sts)
}
\arguments{
\item{sts}{\code{StormsList}}
}
\value{
numeric. The \code{getNbStorms()} returns the number of \code{Storm} objects in the given \code{StormsList} object.
}
\description{
The \code{getNbStorms()} functions allows to get the number of \code{Storm} objects in a \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the number of storms in the sts object
getNbStorms(sts)
}
}
