% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbStorms}
\alias{getNbStorms}
\title{Get number of Storms in StormsList}
\usage{
getNbStorms(sts)
}
\arguments{
\item{sts}{StormsList object}
}
\value{
numeric. Number of storms available in the given StormsList object
}
\description{
Get the number of Storm available in a StormsList object
}
\examples{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
nb <- getNbStomrms(sts)
}
