% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getNbStorms}
\alias{getNbStorms}
\alias{getNbStorms,Storms-method}
\title{Get the number of storms provided in a Storms object}
\usage{
getNbStorms(sts)

\S4method{getNbStorms}{Storms}(sts)
}
\arguments{
\item{sts}{Storms object}
}
\value{
numeric. Number of storms available in the given Storms object
}
\description{
Get the number of storms available in a Storms object
}
