% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getObs}
\alias{getObs}
\alias{getObs,StormsList-method}
\alias{getObs,Storm-method}
\title{Get observations available}
\usage{
getObs(s, ...)

\S4method{getObs}{StormsList}(s, name, season = NULL)

\S4method{getObs}{Storm}(s)
}
\arguments{
\item{s}{\code{Storm}/\code{StormsList} object}

\item{...}{extra argument depending on \code{Storm}/\code{StormsList} object}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the \code{StormsList} object share the same name. Default
value is set to \code{NULL}}
}
\value{
data.frame. Observations of the given storm
}
\description{
Get the observations available for a given storm
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")

## For storm Niran
getObs(getStorm(sts, name = "NIRAN"))

## Equivalent to
getObs(sts, name = "NIRAN")
}
}
