% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getObs}
\alias{getObs}
\alias{getObs,Storms-method}
\alias{getObs,Storm-method}
\title{Get the observations available for a given storm}
\usage{
getObs(s, ...)

\S4method{getObs}{Storms}(s, name, season = NULL)

\S4method{getObs}{Storm}(s)
}
\arguments{
\item{s}{Storms or Storm object}

\item{...}{Additional arguments}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the Storms object share the same name. Default value is set
to NULL}
}
\value{
data.frame. Observations of the given storm
}
\description{
Get the observations available for a given storm
}
