% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getObs}
\alias{getObs}
\alias{getObs,StormsList-method}
\alias{getObs,Storm-method}
\title{Getting observations}
\usage{
getObs(s, ...)

\S4method{getObs}{StormsList}(s, name, season = NULL)

\S4method{getObs}{Storm}(s)
}
\arguments{
\item{s}{\code{Storm} or \code{StormsList} object}

\item{...}{extra argument for \code{StormsList}}

\item{name}{character. Name of the storm in capital letters.}

\item{season}{numeric. Cyclonic season of the \code{Storm}. Required only
if several \code{Storm} in the \code{s} object have the same name. Default value is set
to \code{NULL}.}
}
\value{
data.frame.
}
\description{
The \code{getObs()} function returns observed track data for a storm in a \code{Storm} or \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the observed track data for the tropical cyclone Niran in the sts object 
getObs(getStorm(sts, name = "NIRAN"))
getObs(sts, name = "NIRAN")
}
}
