% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getObs}
\alias{getObs}
\title{Get observations available}
\usage{
getObs(s, ...)
}
\arguments{
\item{s}{Storm/StormsList object}

\item{name}{character. Name of the storm to extract in capital letters}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the StormsList object share the same name. Default value is set
to NULL}
}
\value{
data.frame. Observations of the given storm
}
\description{
Get the observations available for a given storm
}
\examples{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))

## For storm Niran
getObs(getStorm(sts, name = "NIRAN"))

## Equivalent to 
getObs(sts, name = "NIRAN")
}
