% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSeasons}
\alias{getSeasons}
\alias{getSeasons,Storm-method}
\alias{getSeasons,StormsList-method}
\title{Get the cyclonic season(s) of storm(s)}
\usage{
getSeasons(s)

\S4method{getSeasons}{Storm}(s)

\S4method{getSeasons}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm}/\code{StormsList}}
}
\value{
numeric vector. Cyclonic season(s) of each storms provided by the given \code{Storm}
/ \code{StormsList}
}
\description{
Get the cyclonic season(s) of storm(s) available in a \code{Storm} / \code{StormsList}
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")

## For storm Niran
getSeasons(getStorm(sts, name = "NIRAN"))

## For sts
getSeasons(sts)
}
}
