% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSeasons}
\alias{getSeasons}
\alias{getSeasons,Storm-method}
\alias{getSeasons,StormsList-method}
\title{Getting cyclonic seasons of the storms}
\usage{
getSeasons(s)

\S4method{getSeasons}{Storm}(s)

\S4method{getSeasons}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm} or \code{StormsList} object.}
}
\value{
numeric vector.
}
\description{
The \code{getSeasons()} function  returns the cyclonic season of each storm in a \code{Storm} or \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting the cyclonic seasons of the storms from the sts object 
getSeasons(sts)
}
}
