% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSeasons}
\alias{getSeasons}
\alias{getSeasons,Storms-method}
\alias{getSeasons,Storm-method}
\title{Get the cyclonic seasons of storm(s) provided in a Storms/Storm object}
\usage{
getSeasons(s, ...)

\S4method{getSeasons}{Storms}(s, name = NULL)

\S4method{getSeasons}{Storm}(s)
}
\arguments{
\item{s}{Storms or Storm object}

\item{...}{Additional argument}

\item{name}{character. Name of a storm in capital letters. Default value is
set to NULL}
}
\value{
numeric vector. Cyclonic seasons of each storms (or the selected one
through name input) provided by the given Storms object or the cyclonic
season for the given Storm object
}
\description{
Get the cyclonic season of storms available in a Storms object or the
cyclonic season of a Storm object
}
