% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSeasons}
\alias{getSeasons}
\title{Get the cyclonic seasons of storm(s)}
\usage{
getSeasons(s)
}
\arguments{
\item{s}{Storm/StormsList object}
}
\value{
numeric vector. Cyclonic seasons of each storms (or the selected one
through name input) provided by the given StormsList object or the cyclonic
season for the given Storm object
}
\description{
Get the cyclonic season of storms available in a Storm/StormsList
}
\examples{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))

## For storm Niran
getSeasons(getStorm(sts, name = "NIRAN"))

## For sts
getSeasons(sts)
}
