% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSSHS}
\alias{getSSHS}
\alias{getSSHS,Storm-method}
\alias{getSSHS,StormsList-method}
\title{Get maximum SSHS reached of storm(s)}
\usage{
getSSHS(s)

\S4method{getSSHS}{Storm}(s)

\S4method{getSSHS}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm}/\code{StormsList} object}
}
\value{
numeric vector. Maximum SSHS reached of each storms (or the selected
one through \code{name} input) provided by the given \code{StormsList} object or
maximum SSHS reached for the given \code{Storm} object
}
\description{
Get the maximum Saffir Simpson Hurricane Scale (SSHS) reached of storms
available in a \code{Storm}/\code{StormsList} object
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")

## For storm Niran
getSSHS(getStorm(sts, name = "NIRAN"))

## For sts
getSSHS(sts)
}
}
