% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSSHS}
\alias{getSSHS}
\alias{getSSHS,Storm-method}
\alias{getSSHS,StormsList-method}
\title{Get maximum SSHS reached of storm(s)}
\usage{
getSSHS(s)

\S4method{getSSHS}{Storm}(s)

\S4method{getSSHS}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm} / \code{StormsList}}
}
\value{
numeric vector. Maximum SSHS reached for each storm provided by the
given \code{Storm} / \code{StormsList}
}
\description{
Get the maximum Saffir Simpson Hurricane Scale (SSHS) reached of storm(s)
available in a \code{Storm} / \code{StormsList}
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")

## For storm Niran
getSSHS(getStorm(sts, name = "NIRAN"))

## For sts
getSSHS(sts)
}
}
