% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSSHS}
\alias{getSSHS}
\alias{getSSHS,Storm-method}
\alias{getSSHS,StormsList-method}
\title{Getting maximum Saffir-Simpson hurricane wind scale category}
\usage{
getSSHS(s)

\S4method{getSSHS}{Storm}(s)

\S4method{getSSHS}{StormsList}(s)
}
\arguments{
\item{s}{\code{Storm} or \code{StormsList} object.}
}
\value{
numeric vector. The \code{getSSHS()} function return the maximum Saffir-Simpson hurricane wind scale category reached by each storm in the \code{Storm} or \code{StormsList} object.
}
\description{
The \code{getSSHS()} function allows to get the maximum Saffir-Simpson hurricane wind scale category reached by each storm in a \code{Storm} or \code{StormsList} object.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for all storms near New Caledonia
sts <- Storms(sds=sds, loi = "New Caledonia")

#Getting maximum Saffir-Simpson hurricane wind scale category reached by each storm in the sts object 
getSSHS(sts)
}
}
