% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSSHS}
\alias{getSSHS}
\alias{getSSHS,Storms-method}
\alias{getSSHS,Storm-method}
\title{Get the maximum SSHS reached of storm(s) provided in a Storms/Storm object}
\usage{
getSSHS(s, ...)

\S4method{getSSHS}{Storms}(s, name = NULL)

\S4method{getSSHS}{Storm}(s)
}
\arguments{
\item{s}{Storms or Storm object}

\item{...}{Additional argument}

\item{name}{character. Name of a storm in capital letters. Default value is
set to NULL}
}
\value{
numeric vector. Maximum SSHS reached of each storms (or the selected
one through name input) provided by the given Storms object or maximum SSHS
reached for the given Storm object
}
\description{
Get the maximum Saffir Simpson Hurricane Scale (SSHS) reached of storms
available in a Storms object or the maximum SSHS reached of a Storm object
}
