% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getSSHS}
\alias{getSSHS}
\title{Get maximum SSHS reached of storm(s)}
\usage{
getSSHS(s)
}
\arguments{
\item{s}{Storm/StormsList object}

\item{name}{character. Name of a storm in capital letters. Default value is
set to NULL}
}
\value{
numeric vector. Maximum SSHS reached of each storms (or the selected
one through name input) provided by the given StormsList object or maximum SSHS
reached for the given Storm object
}
\description{
Get the maximum Saffir Simpson Hurricane Scale (SSHS) reached of storms
available in a Storm/StormsList object
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))

## For storm Niran
getSSHS(getStorm(sts, name = "NIRAN"))

## For sts
getSSHS(sts)
}
}
