% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getStorm}
\alias{getStorm}
\alias{getStorm,StormsList-method}
\title{Extract a \code{Storm}}
\usage{
getStorm(sts, name, season = NULL)

\S4method{getStorm}{StormsList}(sts, name, season = NULL)
}
\arguments{
\item{sts}{\code{StormsList}}

\item{name}{character. Name of the storm to extract}

\item{season}{numeric. Cyclonic season of the \code{Storm} to extract. Used only
if several \code{Storm} in the \code{sts} share the same name. Default value is set
de \code{NULL}}
}
\value{
\code{Storm}
}
\description{
Extract a \code{Storm} from a \code{StormsList}
}
\examples{
\dontrun{
sts <- Storms(loi = "New Caledonia")
st <- getStorm(sts, name = "NIRAN")
}
}
