% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getStorm}
\alias{getStorm}
\title{Extract a Storm}
\usage{
getStorm(sts, name, season = NULL)
}
\arguments{
\item{sts}{StormsList object}

\item{name}{character. Name of the storm to extract}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the StormsList object share the same name.
Default value is set de NULL}
}
\value{
Storm object
}
\description{
Extract a Storm object from a StormsList object
}
\examples{
sts <- Storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
st <- getStorm(sts, name = "NIRAN")
}
