% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{methods}
\name{getStorm}
\alias{getStorm}
\alias{getStorm,Storms-method}
\title{Get all informations about a storm from a Storms object}
\usage{
getStorm(sts, name, season = NULL)

\S4method{getStorm}{Storms}(sts, name, season = NULL)
}
\arguments{
\item{sts}{Storms object}

\item{name}{character. Name of the storm to extract}

\item{season}{numeric. Cyclonic season of the storm to extract. Used only if
several storms in the Storms object share the same name.
Default value is set de NULL}
}
\value{
Storm object
}
\description{
Extract a Storm object from a Storms object
}
