% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorms}
\alias{getStorms}
\title{Initialize a Storms object}
\usage{
getStorms(
  sdb = IBTRACS,
  loi,
  seasons = c(1980, 2022),
  names = NULL,
  max_dist = 300,
  verbose = FALSE,
  remove_TD = TRUE
)
}
\arguments{
\item{sdb}{list. TC data base loaded from a stormDatabase object}

\item{loi}{Location of Interest. Must be either:
\itemize{
\item a SpatialPolygon (shapefile)
\item a sf object
\item a point of longitude/latitude coordinates provided in a numeric vector
\item a character representing a country,
}}

\item{seasons}{numeric vector. Should be either one or a range of calendar
years. For cyclones that formed in one year and dissipated in the following
year, the latter should be used. It could also be a vector of cyclonic season provided
that it has the same length as names input and matches the season of each storm
listed in names input. Default value is set to c(1980, 2021)}

\item{names}{character vector. Name(s) of storm(s). Default value is set to NULL,
otherwise seasons and names must have the same length, and must match with
a storm on the IBTrACS database}

\item{max_dist}{numeric. Indicates the buffer distance (in km) used to generate
spatial.loi.buffer. Default value is set to 300km. This value also represents
the maximum distance from the track of the storm where computations can be performed}

\item{verbose}{logical. Whether or not the function must be verbose and display
a text progress bar. Default value is set to FALSE}

\item{remove_TD}{logical. Whether or not to remove tropical depressions (< 18 m/s)
and include cyclones only. Default value is set to TRUE.}
}
\value{
a Storms object that gathers all storms that match the criteria given
in the inputs
}
\description{
This function returns a Storms object that
gathers all the storms specified by the user
}
\examples{
#Focus on a single storm
pam <- getStorms(seasons = 2015, names = "PAM", loi = "Vanuatu")

#Focus on several storms over Vanuatu
sts_nc <- getStorms(seasons = c(2003,2021), names = c("ERICA","NIRAN"), loi = "New Caledonia")

#Focus on every storms that occured in the WP basin between 2010 and 2020
poly <- cbind(c(100, 180, 180, 100, 100),c(0, 0, 60, 60, 0))
wp <- sf::st_polygon(list(poly))
wp <- sf::st_sfc(wp, crs = 4326)
wp <- sf::st_as_sf(wp)
sts_wp <- getStorms(loi = wp, seasons = c(2010,2020), verbose = TRUE)

}
