% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorms}
\alias{getStorms}
\title{Initialize a S4 Storms object}
\usage{
getStorms(time_period = c(1970, 2022), name = NULL, loi = "SP", max_dist = 300)
}
\arguments{
\item{time_period}{A numeric vector that contains either the time range of
cyclonic season we are interested in, or the cyclonic seasons that matches
the occurencies of storms contains in \code{name}. Default value is set to
c(1970,2020)}

\item{name}{A character vector that contains the name of the storms we are
interested in. It must be the same size of \code{time_period}. If the time_period
do not correspond, the function will throw an error}

\item{loi}{Location of interest. Should be either a \code{SpatialPolygon}, or a sf
object. Default value is set to "SP" which will focus on the whole South
West Pacific Basin. It can also be a numeric vector that contains one longitude/
latitude coordinate. If so, a circle based \code{SpatialPolygon}
centered in loi, with a radius of \code{max_dist} is used as loi.}

\item{max_dist}{Numeric that indicates the radius (in km) of the
buffer to generate \code{spatail.poly.buffer}. Default value is set to 300km.}
}
\value{
a S4 Storms object that gathers all the above informations
}
\description{
Initialize a S4 Storms object
}
