% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorms}
\alias{getStorms}
\title{Initialize a Storms object}
\usage{
getStorms(
  basin = "SP",
  time_period = c(1980, 2022),
  name = NULL,
  loi = NULL,
  max_dist = 300,
  verbose = FALSE,
  remove_TD = TRUE
)
}
\arguments{
\item{basin}{character. Name of basin where the Storms should be extracted.
Default value is set to "SP"}

\item{time_period}{numeric vector. Should be either one cyclonic season or a range
of cyclonic season. It could also be a vector of cyclonic season provided
that it has the same length as name and matches the season of each Storm
listed in name. Default value is set to c(1980, 2021)}

\item{name}{character vector. Name(s) of storm(s). Default value is set to NULL,
otherwise time_period and name must have the same length, and these two
informations must match}

\item{loi}{Location of Interest. Should be either a SpatialPolygon, a sf
object, a point of coordinates in lon/lat, a character representing a country,
or a basin. Default value is set to NULL which will set the spatial.loi.buffer
on the whole basin}

\item{max_dist}{numeric. Indicates the buffer used to generate
spatail.loi.buffer (in km). Default value is set to 300}

\item{verbose}{logical. Whether or not the function must be verbose and display
a text progress bar. Default value is set to FALSE}

\item{remove_TD}{logical. Whether or not to remove Tropical Depression (< 18 m/s).
Default value is set to TRUE.}
}
\value{
a Storms object that gathers all storms that match the criteria given
in the inputs
}
\description{
Depending on values of the inputs, this function return a Storms object that
gathers all the storms and tropical cyclone the user is interested in
}
\examples{
#Focus on a single storm
pam = getStorms(time_period = 2015, name = "PAM", loi = "Vanuatu")

#Focus on several storms over Vanuatu
sts_nc = getStorms(time_period = c(2003,2021), name = c("ERICA","NIRAN"), loi = "New Caledonia")

#Focus on every storms that occured in the WP basin between 2010 and 2020
sts_wp = getStorms(basin = "WP", time_period = c(2010,2020), verbose = TRUE)

}
