% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{getStorms}
\alias{getStorms}
\title{Initialize a Storms object depending on a selection}
\usage{
getStorms(
  basin = "SP",
  time_period = c(1980, 2022),
  name = NULL,
  loi = NULL,
  max_dist = 300,
  verbose = FALSE,
  remove_weak_TC = TRUE
)
}
\arguments{
\item{basin}{character. Name of basin where the Storms should be extracted.
Default value is set to \code{"SP"}}

\item{time_period}{numerics. Should be either one cyclonic season or a range
of cyclonic season. It could also be a vector of cyclonic season provided
that it has the same length as \code{name} and matches the season of each Storm
listed in \code{name}. Default value is set to c(1980, 2021)}

\item{name}{characters. Name(s) of Storm(s). Default value is set to NULL,
otherwise \code{time_period} and \code{name} must have the same length, and these two
informations must match}

\item{loi}{Location of Interest. Should be either a \code{SpatialPolygon}, a \code{sf}
object, a point of coordinates in lon/lat, a character representing a country,
or a basin. Default value is set to \code{NULL} which will set the \code{spatial.loi.}
on the whole \code{basin}}

\item{max_dist}{numeric. Indicates the buffer used buffer to generate
\code{spatail.loi.buffer} (in km). Default value is set to 300}

\item{verbose}{logical. Whether or not the function must be verbose and display
a text progress bar. Default value is set to \code{FALSE}}

\item{remove_weak_TC}{logical. Whether or not to remove Stomrs under category
1 in the Saffir Simpson Hurricane Scale. Default value is set to TRUE.}
}
\value{
a S4 Storms object that gathers all the above informations
}
\description{
Initialize a Storms object depending on a selection
}
