% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{Holland80_profile}
\alias{Holland80_profile}
\title{Compute wind values according to the Holland 80 model}
\usage{
Holland80_profile(r, rmw, wmax, pres, poci, lat)
}
\arguments{
\item{r}{Distance to the center of the storm where the value must be computed}

\item{rmw}{Radius of maximum wind speed}

\item{pres}{Pressure at the center of the storm}

\item{poci}{Pressure at the Outermost Closed Isobar}

\item{lat}{Latitude of the storm (eye)}

\item{b}{Scaling factor}
}
\value{
wind value according to the Holland 80 model at distance \code{r} to the
center of the storm
}
\description{
Compute wind values according to the Holland 80 model
}
