% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IBTRACS_SP}
\alias{IBTRACS_SP}
\title{South Pacific storms database}
\format{
\subsection{\code{IBTRACS_SP}}{

a StormsDataset object
}
}
\usage{
IBTRACS_SP
}
\description{
Toy database extracted from the IBTrACS.SP.v04r00.nc file.
Provides all the tropical cyclones that occured around Vanuatu from 2015 to 2016
and around New Caledonia from 2020 to 2021
}
\keyword{datasets}
