% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsDataset.R
\name{initDatabase}
\alias{initDatabase}
\title{Initialize a StormsDataset object}
\usage{
initDatabase(
  filename = system.file("extdata", "IBTrACS.SP.v04r00.nc", package = "StormR"),
  fields = c(basin = "basin", names = "name", seasons = "season", isoTime = "iso_time",
    lon = "usa_lon", lat = "usa_lat", msw = "usa_wind", rmw = "usa_rmw", pressure =
    "usa_pres", poci = "usa_poci", sshs = "usa_sshs"),
  basin = "SP",
  verbose = TRUE
)
}
\arguments{
\item{filename}{character. Name of the file to be loaded}

\item{fields}{named character vector. Dictionary that provides all the name of dimension to extract from the netcdf database.
See Details}

\item{basin}{character. Basin name to filter the database within its boundaries. If NULL, no filter is performed and the
whole database will be collected when calling collectData method.}

\item{verbose}{logical.}
}
\value{
An object of class StormsDataset
}
\description{
Initialize a StormsDataset object
}
