% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_storm_track.R
\name{interpolate_storm_track}
\alias{interpolate_storm_track}
\title{Interpolate Storm Track}
\usage{
interpolate_storm_track(example_infile)
}
\arguments{
\item{example_infile}{Describe Input}
}
\value{
Describe Output
}
\description{
This function interpolates a storm track to 30 minute increments.
}
