% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made_landfall_US.R
\name{made_landfall_US}
\alias{made_landfall_US}
\title{Determine if the Storm Made Landfall in the United States}
\usage{
made_landfall_US(df)
}
\arguments{
\item{df}{A dataframe containing the storm data for a single storm}
}
\value{
TRUE if the storm made landfall in the continental United States, FALSE otherwise
}
\description{
A function for determining whether the storm made landfall in the
continental United States based on its latitude and longitude values.
}
