% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam}
\alias{pam}
\title{PAM tropical cyclone}
\format{
\subsection{\code{pam}}{

a Storms object generated with getStorms
}
}
\usage{
pam
}
\description{
Gathers informations about the category 5 TC PAM (2015) over Vanuatu with a
buffer of 300km
}
\keyword{datasets}
