% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plot_labels}
\alias{plot_labels}
\title{Add ISO_time and name label for the first and the
last observation within the loi, on a map for a given storm.}
\usage{
plot_labels(storm)
}
\arguments{
\item{storm}{Storm object that we want to plot the track. Note that both a map
and track of the storm should be previously plotted}
}
\description{
Add ISO_time and name label for the first and the
last observation within the loi, on a map for a given storm.
}
