% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plot_labels}
\alias{plot_labels}
\title{Add ISO_time and name label for the first and the
last observation within the loi, on a map for a given storm.}
\usage{
plot_labels(storm, by, pos)
}
\arguments{
\item{storm}{Storm object that we want to plot the track. Note that both a map
and track of the storm should be previously plotted}

\item{by}{number, increment of the sequence for the labels to plot}

\item{pos}{number between 1 and 4, position for the labels to plot: up, left,
down, right
\code{spatial.loi.buffer} of object \code{sts}}
}
\description{
Add ISO_time and name label for the first and the
last observation within the loi, on a map for a given storm.
}
