% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBehaviour.R
\name{plotBehaviour}
\alias{plotBehaviour}
\title{Plot a rasterize product (Maximum Sustained Wind, Power Dissipation Index,
Category Exposure, 2D wind speed structure at a given observation ...)
associated with a Storm contained in a \code{Storms} alongside with its track}
\usage{
plotBehaviour(sts, raster_product, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{sts}{\code{Storms} object}

\item{raster_product}{\code{Spatraster} object. Name of the layer must be
"stormName_product" where product is either MSW, PDI, Exposure(1,2,3,4,5,All).
It can also be "stormName_profileInd" where Ind stand for the observations if
\code{raster_product} is a 2D wind speed structure}

\item{xlim}{numerics. A set of longitude coordinates that controls the
longitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the x bounding box of \code{spatial.loi.buffer}.}

\item{ylim}{numerics. A set of latitude coordinates that controls the
latitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the y bounding box of \code{spatial.loi.buffer}.}
}
\description{
Plot a rasterize product (Maximum Sustained Wind, Power Dissipation Index,
Category Exposure, 2D wind speed structure at a given observation ...)
associated with a Storm contained in a \code{Storms} alongside with its track
}
