% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBehaviour.R
\name{plotBehaviour}
\alias{plotBehaviour}
\title{Plot a selected product (MSW, PDI ...) alongside with the track of a selected
storm contained in object \code{sts}.}
\usage{
plotBehaviour(sts, raster_product, xlim = NULL, ylim = NULL, mask = FALSE)
}
\arguments{
\item{sts}{Storms object that contains the storm we are interested in}

\item{raster_product}{Spatraster object which spatializes the product we are interested
in. The name of the layer must be nameOfTheStorm_product (which is the case)
if the product is generated with the function \code{stormBehaviour}}

\item{xlim}{A set of longitude coordinates that controls the longitude extent
of the plot. Default value is NULL which will let the plot extends according
to the x bounding box of \code{spatial.loi.buffer}.}

\item{ylim}{A set of latitude coordinates that controls the latitude extent
of the plot. Default value is NULL which will let the plot extends according
to the y bounding box of \code{spatial.loi.buffer}.}

\item{mask}{Logical, wether or not to mask the data according to \code{spatial.loi}}
}
\description{
Plot a selected product (MSW, PDI ...) alongside with the track of a selected
storm contained in object \code{sts}.
}
