% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBehaviour.R
\name{plotBehaviour}
\alias{plotBehaviour}
\title{Plot rasterize informations below the associated track of a storm}
\usage{
plotBehaviour(
  sts,
  raster_product,
  xlim = NULL,
  ylim = NULL,
  labels = FALSE,
  by = 8,
  pos = 3
)
}
\arguments{
\item{sts}{\code{Storms} object}

\item{raster_product}{\code{Spatraster} object. Name of the layer must be
"stormName_product" where product is either MSW, PDI, Exposure(1,2,3,4,5,All).
It can also be "stormName_profileInd" where Ind stand for the observations if
\code{raster_product} is a 2D wind speed structure}

\item{xlim}{numeric vector. A set of longitude coordinates that controls the
longitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the x bounding box of \code{spatial.loi.buffer}.}

\item{ylim}{numeric vector. A set of latitude coordinates that controls the
latitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the y bounding box of \code{spatial.loi.buffer}.}

\item{labels}{logical. Whether or not to plot ISO Times and name labels}

\item{by}{numeric. Increment of the sequence for the labels to plot. Default value
is set to 8 which represents a 24h time interval}

\item{pos}{numeric. Must be between 1 and 4 and correspond to the position of
labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right).
Default value is set to 3}
}
\description{
This function plots a rasterize product (Maximum Sustained Wind, Power Dissipation Index,
Category Exposure, 2D wind speed structure at a given observation ...)
associated with a storm contained in a \code{Storms} object alongside with its track
}
