% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBehaviour.R
\name{plotBehaviour}
\alias{plotBehaviour}
\title{Plot rasterized information storm behaviour}
\usage{
plotBehaviour(
  sts,
  raster_product,
  color_palette = NULL,
  xlim = NULL,
  ylim = NULL,
  labels = FALSE,
  by = 8,
  pos = 3
)
}
\arguments{
\item{sts}{Storms object}

\item{raster_product}{Spatraster object. Name of the layer must be
"stormName_product" where product is either "MSW", "PDI", "Exposure_min-max",
"Profiles_index", "WindDirection_index" where "min" and "max" represent the
threshold used to compute Exposure raster and index stands for the index of observation.}

\item{color_palette}{character vector. Represents the color palette used for the plot.
Default value is set to NULL, which will automatically choose a color palette
provided by this package and depending on the product}

\item{xlim}{numeric vector. A set of longitude coordinates that controls the
longitude extent of the plot. Default value is set to NULL which will let
the plot extends according to the x bounding box of the spatial LOI buffer
provided in sts}

\item{ylim}{numeric vector. A set of latitude coordinates that controls the
latitude extent of the plot. Default value is set to NULL which will let
the plot extends according to the y bounding box of the spatial LOI buffer
provided in sts}

\item{labels}{logical. Whether or not to plot ISO Times and name labels}

\item{by}{numeric. Defines the frequency at which labels are plotted for the
3-hourly records. Default value is set to 8 which represents a 24h time interval
between each labeled observations. Ignored if labels == FALSE}

\item{pos}{numeric. Must be between 1 and 4. Correspond to the position of
labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right).
Default value is set to 3. Ignored if labels == FALSE}
}
\description{
This function plots a rasterize product (maximum sustained wind, power
dissipation index, category exposure, or 2D wind speed structure and Wind direction at a given
observation) associated with a storm contained in a Storms object
alongside with its track
}
\examples{
#Plot MSW analytic raster for Pam (2015) in Vanuatu
pam <- getStorms(loi = "Vanuatu", names = "PAM")
pam_msw <- stormBehaviour_sp(pam)
plotBehaviour(pam, pam_msw)

#Plot a 2D windspeed structure  for Pam (2015) in Vanuatu
pam <- getStorms(loi = "Vanuatu", names = "PAM")
pam_prof <- stormBehaviour_sp(pam, format = "profiles", verbose = 3)
plotBehaviour(pam, pam_prof[["PAM_Profiles_37"]], labels = TRUE, pos = 2)


}
