% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotStorms}
\alias{plotStorms}
\title{Plot several storm tracks}
\usage{
plotStorms(
  sts,
  names = NULL,
  category = NULL,
  map = NULL,
  ground_color = "grey",
  ocean_color = "white",
  whole_basin = FALSE,
  labels = FALSE,
  by = 8,
  pos = 3,
  legends = FALSE,
  loi = TRUE,
  grtc = 1,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{sts}{Storms object}

\item{names}{character vector. Names of the Storms we would like to plot. Default
value is set to \code{NULL} which will consider every Storm in \code{sts}}

\item{category}{numeric vector. Should be either a category or a range of category
in the Saffir Simpson scale (-2 to 5). Default value is set to \code{NULL} which
will consider every Storm in \code{sts}. Otherwise it will consider only storm that
reached \code{category}}

\item{map}{a \code{shapefile} or \code{sf} object. It should replace
the default map. Default value is set to NULL.}

\item{ground_color}{character. Color for the ground}

\item{ocean_color}{character. Color for the oceans}

\item{whole_basin}{logical. Whether or not to plot the track onto the whole basin.
Default value is set to \code{FALSE}. Otherwise, the plot focuses on the extent of
\code{spatial.loi.buffer} of \code{sts}}

\item{labels}{logical. Whether or not to plot ISO Times and name labels}

\item{by}{numeric. Increment of the sequence for the labels to plot. Default value
is set to 8 which represents a 24h time interval}

\item{pos}{numeric. Must be between 1 and 4 and correspond to the position of
labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right).
Default value is set to 3}

\item{legends}{logical. Whether or not to plot legends. Default value is set
to FALSE.}

\item{loi}{logical. Whether or not to plot \code{spatial.loi.buffer} on the map
Default value is set to TRUE.}

\item{grtc}{numeric. Controls the number of graticules to plot. Default value
is set to 1, which plots graticules on multiple of 10 coordinates. Note that
it should be a power of 2.}

\item{xlim}{numeric vector. A set of longitude coordinates that controls the
longitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the x bounding box of \code{spatial.loi.buffer}.}

\item{ylim}{numeric vector. A set of latitude coordinates that controls the
latitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the x bounding box of \code{spatial.loi.buffer}.}
}
\description{
This function plots a set of storm tracks contained in a Storms object. Depending
on the inputs, the user may choose to plot only a desired set of storm extracted
from the Storms object.
}
