% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotStorms}
\alias{plotStorms}
\title{Plot storm track(s)}
\usage{
plotStorms(
  sts,
  names = NULL,
  category = NULL,
  xlim = NULL,
  ylim = NULL,
  labels = FALSE,
  by = 8,
  pos = 3,
  legends = "topleft",
  loi = TRUE
)
}
\arguments{
\item{sts}{\code{StormsList}}

\item{names}{character vector. Name(s) of the storm(s) in capital letters to
plot on the map. Default value is set to \code{NULL}, which will plot every
storm provided in \code{sts}. To see which \code{Storm} are included in a \code{StormsList}
, use the \code{getNames} function}

\item{category}{numeric vector. Should be either a category or a range of
categories in the Saffir Simpson scale (-1 to 5). Default value is set to
\code{NULL} which will consider every \code{Storm} in \code{sts}. Otherwise it will consider
only \code{Storm} that reached \code{category} input}

\item{xlim}{numeric vector. A set of longitude coordinates that controls the
longitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the longitude range of the extended LOI}

\item{ylim}{numeric vector. A set of latitude coordinates that controls the
latitude extent of the plot. Default value is set to \code{NULL} which will let
the plot extends according to the the latitude range of the extended LOI}

\item{labels}{logical. Whether or not to plot names labels with the
corresponding indices of observations and ISO Times along the track(s).
Default value is set to \code{NULL}}

\item{by}{numeric. Defines the frequency at which labels are plotted for the
3 (or 6) hourly records. Default value is set to \code{8} which represents a 24h
(or 48h) time interval between each labeled observations. Ignored if
\code{labels == FALSE}}

\item{pos}{numeric. Must be between \code{1} and \code{4}. Correspond to the position
of labels according to the observation: \code{1} (up), \code{2} (left), \code{3} (down),
\code{4} (right). Default value is set to 3. Ignored if \code{labels == FALSE}}

\item{legends}{character. Indicates the where the legend should be plotted.
Must be either \code{"topright"}, \code{"topleft"},  (default setting),
\code{"bottomleft"}, \code{"bottomright"} or \code{"none"} (no legend)}

\item{loi}{logical. Whether or not to plot the extended LOI on the map.
Default value is set to \code{TRUE}}
}
\description{
This function plots a set of storm tracks contained in a \code{StormsList}
(See \code{Storms} function). Depending on the inputs, the user can choose to plot
only a desired set of \code{Storm} extracted from the \code{StormsList}
}
\examples{
\dontrun{
#' #Create database (Default)
sds <- defDatabase()

#Get Pam tropical cyclone (2015)
pam <- Storms(sds = sds, loi = "Vanuatu", names = "PAM")

#Plot Pam over Vanuatu with labels every 24h
plotStorms(pam, labels = TRUE)

#Plot Pam, with labels every 6h on the right side of observations
plotStorms(pam, labels = TRUE, by = 2, pos = 4)
}


}
