% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotStorms}
\alias{plotStorms}
\title{Plot a set of storm}
\usage{
plotStorms(
  sts,
  shapefile = NULL,
  ground_color = "grey",
  ocean_color = "white",
  all_basin = FALSE,
  labels = FALSE,
  grtc = 1
)
}
\arguments{
\item{sts}{S4 Storms object that gathers all the storms we are interested in}

\item{shapefile}{a \code{shapefile} or \code{SpatialPolygons} object that should replace
the default map. Default value is set to NULL.}

\item{ground_color}{character that stands for the color of the ground area}

\item{ocean_color}{character that stands for the color of the ocean area}

\item{all_basin}{Logical, wether or not to plot track onto the whole basin.
Default value is set to FALSE. Otherwise, the plot focuses on the extent of
\code{spatial.loi} of object \code{sts}}

\item{labels}{Logical, wether or not to plot ISO_time and name labels for the
first and the last observation of each storms within the loi. Default value
is set to FALSE.}

\item{grtc}{numeric that controls how many graticules to plot. Default value
is set to 1, which plots graticules on multipule of 10 coordinates. Note that
it should be a power of 2.}
}
\description{
Plot a set of storm
}
