% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotStorms}
\alias{plotStorms}
\title{Plot a set of storm}
\usage{
plotStorms(
  sts,
  name = NULL,
  shapefile = NULL,
  ground_color = "grey",
  ocean_color = "white",
  all_basin = FALSE,
  labels = FALSE,
  by = 8,
  pos = 3,
  legends = FALSE,
  loi = TRUE,
  grtc = 1,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{sts}{S4 Storms object that gathers all the storms we are interested in}

\item{name}{The storm we would like to plot. Default value is NULL that
will plot all the storms contained in \code{sts}.}

\item{shapefile}{a \code{shapefile} or \code{SpatialPolygons} object that should replace
the default map. Default value is set to NULL.}

\item{ground_color}{character that stands for the color of the ground area}

\item{ocean_color}{character that stands for the color of the ocean area}

\item{all_basin}{Logical, wether or not to plot track onto the whole basin.
Default value is set to FALSE. Otherwise, the plot focuses on the extent of
\code{spatial.loi.buffer} of object \code{sts}}

\item{labels}{Logical, wether or not to plot ISO_time and name labels for the
first and the last observation of each storms within the loi. Default value
is set to FALSE.}

\item{by}{number, increment of the sequence for the labels to plot. Default value
is set to 8 which represents a 24h time interval}

\item{pos}{number between 1 and 4, position for the labels to plot: up, left,
down, right}

\item{legends}{Logical, wether or not to plot legends for the
plot. Default value is set to FALSE.}

\item{loi}{Logical, wether or not to plot \code{spatial.loi.buffer} on the map
Default value is set to TRUE.}

\item{grtc}{numeric that controls how many graticules to plot. Default value
is set to 1, which plots graticules on multipule of 10 coordinates. Note that
it should be a power of 2.}

\item{xlim}{A set of longitude coordinates that controls the longitude extent
of the plot. Default value is NULL which will let the plot extends according
to the x bounding box of \code{spatial.loi.buffer}.}

\item{ylim}{A set of latitude coordinates that controls the latitude extent
of the plot. Default value is NULL which will let the plot extends according
to the y bounding box of \code{spatial.loi.buffer}.}
}
\description{
Plot a set of storm
}
