% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotStorms}
\alias{plotStorms}
\title{Plot storm track(s)}
\usage{
plotStorms(
  sts,
  names = NULL,
  category = NULL,
  xlim = NULL,
  ylim = NULL,
  labels = FALSE,
  by = 8,
  pos = 3,
  legends = TRUE,
  loi = TRUE,
  reset_setting = TRUE
)
}
\arguments{
\item{sts}{StormsList object}

\item{names}{character vector. Name(s) of the storm(s) in capital letters to
plot on the map. Default value is set to NULL, which will plot every storm
provided in sts. To see which storms are included in a StormsList object,
you can use the getter "getNames()"}

\item{category}{numeric vector. Should be either a category or a range of
categories in the Saffir Simpson scale (-1 to 5). Default value is set to
NULL which will consider every storm in sts. Otherwise it will consider only
storms that reached category input}

\item{xlim}{numeric vector. A set of longitude coordinates that controls the
longitude extent of the plot. Default value is set to NULL which will let
the plot extends according to the longitude range of the extended LOI}

\item{ylim}{numeric vector. A set of latitude coordinates that controls the
latitude extent of the plot. Default value is set to NULL which will let
the plot extends according to the the latitude range of the extended LOI}

\item{labels}{logical. Whether or not to plot names labels with the
corresponding indices of observations and ISO Times along the track(s)}

\item{by}{numeric. Defines the frequency at which labels are plotted for the
3 (or 6) hourly records. Default value is set to 8 which represents a 24h
(or 48h) time interval between each labeled observations. Ignored if
labels == FALSE}

\item{pos}{numeric. Must be between 1 and 4. Correspond to the position of
labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right).
Default value is set to 3. Ignored if labels == FALSE}

\item{legends}{logical. Whether or not to plot legends. Default value is set
to TRUE}

\item{loi}{logical. Whether or not to plot the extended LOI on the map.
Default value is set to TRUE}

\item{reset_setting}{logical. Whether the graphical parameter should be reset
on exit. Default value is set to TRUE. It is useful for the plotBehaviour
function. We highly recommend not to change this input}
}
\description{
This function plots a set of storm tracks contained in a StormsList object
(See “Storms” function). Depending on the inputs, the user can choose to
plot only a desired set of storms extracted from the StormsList object
}
\examples{
\dontrun{
#Plot Erica over New Caledonia with labels every 24h
plotStorms(sts_nc, names = "ERICA", labels = TRUE)

#Plot Erica, with labels every 6h on the right side of observations
plotStorms(sts_nc, names = "ERICA", labels = TRUE, by = 2, pos = 4)

}


}
