% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorms.R
\name{plotStorms}
\alias{plotStorms}
\title{Plot several storm tracks}
\usage{
plotStorms(
  sts,
  names = NULL,
  category = NULL,
  map = NULL,
  whole_basin = FALSE,
  labels = FALSE,
  by = 8,
  pos = 3,
  legends = TRUE,
  loi = TRUE,
  xlim = NULL,
  ylim = NULL,
  reset_setting = TRUE
)
}
\arguments{
\item{sts}{Storms object}

\item{names}{character vector. Name(s) of the storm(s) to plot on map. Default
value is set to NULL which will consider every storm in sts}

\item{category}{numeric vector. Should be either a category or a range of category
in the Saffir Simpson scale (-2 to 5). Default value is set to NULL which
will consider every storm in sts. Otherwise it will consider only storms that
reached category input}

\item{map}{a shapefile or sf object. It should replace the default map if non NULL.
Default value is set to NULL.}

\item{whole_basin}{logical. Whether or not to plot the track onto the whole basin.
Default value is set to FALSE. Otherwise, the plot focuses on the extent of
spatial.loi.buffer of sts}

\item{labels}{logical. Whether or not to plot ISO Times and name labels}

\item{by}{numeric. Add labels every by observations. Default value
is set to 8 which represents a 24h time interval between each labeled observations.
Ignored if labels == FALSE}

\item{pos}{numeric. Must be between 1 and 4. Correspond to the position of
labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right).
Default value is set to 3. Ignored if labels == FALSE}

\item{legends}{logical. Whether or not to plot legends. Default value is set
to TRUE}

\item{loi}{logical. Whether or not to plot spatial.loi.buffer on the map
Default value is set to TRUE.}

\item{xlim}{numeric vector. A set of longitude coordinates that controls the
longitude extent of the plot. Default value is set to NULL which will let
the plot extends according to the x bounding box of spatial.loi.buffer from sts input.
Ignored if whole_basin is not NULL}

\item{ylim}{numeric vector. A set of latitude coordinates that controls the
latitude extent of the plot. Default value is set to NULL which will let
the plot extends according to the x bounding box of spatial.loi.buffer from sts input.
Ignored if whole_basin is not NULL}

\item{reset_setting}{logical. Whether the graphical parameter should be reset on exit. Default
value is set to TRUE. It is usefull for the plotBehaviour function. We highly recommand not to change this
input.}
}
\description{
This function plots a set of storm tracks contained in a Storms object. Depending
on the inputs, the user may choose to plot only a desired set of storm extracted
from the Storms object.
}
\examples{
#Plot category 5 TCs in the WP Basin between 2010 and 2020
plotStorms(sts_wp, category = c(3,5))

#Plot a single storm (ERICA) with labels every 24h
plotStorms(sts_nc, names = "ERICA", labels = TRUE)

#Plot a single storm (ERICA), with labels every 6h on the right side
plotStorms(sts_nc, names = "ERICA", labels = TRUE, by = 2, pos = 4)

#Plot a single storm (ERICA)
plotStorms(sts_nc, names = "ERICA")


}
