% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defStormsList.R
\docType{methods}
\name{show,storm-method}
\alias{show,storm-method}
\alias{show,stormsList-method}
\title{Show a \code{storm}/\code{stormsList}}
\usage{
\S4method{show}{storm}(object)

\S4method{show}{stormsList}(object)
}
\arguments{
\item{object}{\code{storm}/\code{stormsList} object}
}
\value{
NULL, only display information about the object
}
\description{
Display the \code{storm}/\code{stormsList} object
}
\examples{
\donttest{
sds <- defStormsDataset()
sts <- defStormsList(sds = sds, loi = "New Caledonia", names = c("LUCAS","NIRAN"))
## Display information about Niran in sts
getStorm(sts, name = "NIRAN")

## Display information about sts
sts
}
}
\keyword{internal}
