% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{spatialBehaviour}
\alias{spatialBehaviour}
\title{Compute indicators of storm behaviour}
\usage{
spatialBehaviour(
  sts,
  product = "MSW",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Chen",
  empirical_rmw = FALSE,
  space_res = "2.5min",
  temp_res = 1,
  verbose = 2
)
}
\arguments{
\item{sts}{\code{StormsList}}

\item{product}{character. Product to compute among:
\itemize{
\item \code{"MSW"}: Maximum Sustained Wind
\item \code{"PDI"}: Power Dissipation Index
\item \code{"Exposure"}: hour exposition for wind greater than \code{wind_threshold}
\item \code{"Profiles"}, 2D wind speed structures of wind speed with wind
direction for each observation
}
Default value is set to \code{"MSW"}}

\item{wind_threshold}{numeric vector. Minimal wind threshold(s) (m/s) to
compute \code{"Exposure"} product. Ignored if \code{"Exposure"} is not part of the
products to compute. Default value is set to Saffir Simpson Hurricane Scale
thresholds}

\item{method}{character. Cyclonic model used to compute \code{product}. Must be
either:
\itemize{
\item \code{"Willoughby"}: model based on fits performed on cyclonic observations
\item \code{"Holland"}: model based on both basic cyclonic Physics and parameters
fitting according to cyclonic observations
\item \code{"Boose"}: asymmetric model based on Holland model
}
Default value is set to \code{"Willoughby"} (See \code{Details})}

\item{asymmetry}{character. Indicates which version of asymmetry to use in
the computations (see \code{Details}). Must be either:
\itemize{
\item \code{"Miyazaki"}: based on the formula derived in Miyazaki et al. (1962)
\item \code{"Chen"}:  based on the formula derived in Chen (1994)
\item \code{"None"}: no asymmetry is added
}
Default value is set to \code{"Chen"}. Ignored if \code{method == "Boose"}}

\item{empirical_rmw}{logical. Whether to compute the radius of maximum wind
empirically or using the radius of maximum wind from the observations.
Default value is set to \code{FALSE}. If \code{TRUE}, a formula extracted from
Willoughby et al. 2006 is used to compute rmw}

\item{space_res}{character. Space resolution for the raster(s) to compute.
Either \code{"30sec"}, \code{"2.5min"}, \code{"5min"} or \code{"10min"}. Default value is set
to \code{"2.5min"}}

\item{temp_res}{numeric. Period of time used to interpolate data. Allowed
values are \code{1} (60min), \code{0.75} (45min), \code{0.5} (30min), and \code{0.25} (15min).
Default value is set to \code{1}}

\item{verbose}{numeric. Whether or not the function should display
informations about the process and/or outputs. Allowed values are:
\itemize{
\item \code{0}: Nothing is displayed
\item \code{1}: Informations about the process are displayed
\item \code{2}: Outputs are also displayed
}
Default value is set to \code{2}}
}
\value{
SpatRaster stack which provides the desired product computed,
projected in WGS84 and spanning over the extented LOI of the \code{StormsList} .
Number of layers depends on the number of \code{Storm} available in \code{sts} and
also \code{product} and \code{temp_res} inputs:
\itemize{
\item \code{"MSW"} produces one layer per \code{Storm}. Name of layer is "STORMNAME_MSW"
\item \code{"PDI"} produces one layer per \code{Storm}. Name of layer is "STORMNAME_PDI"
\item \code{"Exposure"} produces one layer for each wind values available
in \code{wind_threshold} and for each \code{Storm}. Name of layers are
"STORMNAME_Exposure_threshold1", "STORMNAME_Exposure_threshold2"...
\item \code{"Profiles"} produces two layers for each observations
(real and interpolated) and each  \code{Storm}. Name of layers are
"STORMNAME_Speed_observation", "STORMNAME_Direction_observation"
}
}
\description{
This function computes/rasterizes products for each \code{Storm} included in a
\code{StormsList}, among Maximum Sustained Wind, Power Dissipation Index, Category
exposure and 2D wind speed structures/direction of wind speed for every
observations
}
\details{
Add details on Willoughy/Holland/Boose method, asymmetries and
getRMW method ... The \code{temp_res} input will perform a linear interpolation
of observations to further compute each 2D wind speed structure at each
interpolated observations. For example, if \code{temp_res == 1}, it will
generate observations every 1hour between the available observations. Doing
so, 2D wind speed structure is computed and stacked for each observations
(available and interpolated) to compute the desired product(s) afterwards.
If \code{product == "Profiles"}, nothing else is performed. Otherwise and
depending on the product(s) to compute, calculations are carried out on the
stack and \code{terra::focal} functions are applied to the final raster(s) to
smooth the result(s).
}
\examples{
\dontrun{
#Compute MSW product for Pam 2015 in Vanuatu using default settings
msw.pam <- spatialBehaviour(pam)

#Compute PDI product for Erica and Niran in New Caledonia using Holland model without asymmetry
pdi.nc <- spatialBehaviour(sts_nc, method = "Holland", product = "PDI", asymmetry = "None")

#Compute Exposure for Pam 2015 in Vanuatu using default settings
exp.pam <- spatialBehaviour(pam, product = "Exposure")

#Compute profiles wind speed for Erica and Niran in New Caledonia using default settings
prof.nc <- spatialBehaviour(sts_nc, product = "Profiles")
}

}
