% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{Storm-class}
\alias{Storm-class}
\alias{Storm}
\title{Gather all the informations to model a single Storm}
\value{
A S4 object gathering all the above informations
}
\description{
Gather all the informations to model a single Storm
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the storm}

\item{\code{season}}{numeric. Cyclonic season in which the storm has occured}

\item{\code{basin}}{character. Basin in which the storm has occured}

\item{\code{sshs}}{numeric. Category in the Saffir Simpson Hurricane Scale}

\item{\code{numobs.all}}{numeric. Total number of observations available.}

\item{\code{obs.all}}{data.frame. Contains all of the observations available.
An observation is made up of several slots which are:
Basin, subbasin, iso.time, lon, lat, msw (Maximum Sustained Wind in), rmw (Radius of Maximum Wind),
roci (Radius of the Outermost Closed Isobar), pres (pressure at the center),
poci (Pressure of the Outermost Closed Isobar), sshs (Category in the Saffir Simpson Hurricane Scale),
landfall (Minimum distance to land over next 3 hours,  = 0 means landfall)).}

\item{\code{numobs}}{numeric. Total number of observations available within the location of interest + buffer}

\item{\code{obs}}{numerics. Indices of observations within the location of interest + buffer}

\item{\code{lty.track}}{numeric. Indicates which line type is used to plot the storm}
}}

