% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{Storm-class}
\alias{Storm-class}
\alias{Storm}
\title{Models a storm using various slots}
\value{
A S4 object gathering all the above informations
}
\description{
Models a storm using various slots
}
\section{Slots}{

\describe{
\item{\code{name}}{A character that contains the name of the storm}

\item{\code{sid}}{A character that contains a unique id given by ibtracs algorithm}

\item{\code{season}}{A numeric that contains the cyclonic season in which the storm has occured.}

\item{\code{numobs.all}}{A numeric that contains the total number of observations available.}

\item{\code{obs.all}}{A data.frame.that contains of the observations available, where an
observation is: Basin, Subbasin, ISO_time, lon, lat, wmo_msw, Nadi_wind, Nadi_cat}

\item{\code{numobs}}{A numeric that contains the number of observations available within
the area of interest/location of interest}

\item{\code{obs}}{A data.frame.that contains of the observations available within
the area of interest/location of interest}

\item{\code{first.obs}}{A numeric standing for the last observation before entering
the area of interest/location of interest}

\item{\code{last.obs}}{A numeric standing for the first observation after leaving
the area of interest/location of interest}

\item{\code{lty.track}}{A numeric that indicates which line type should be used to
plot the storm}
}}

