% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{Storm-class}
\alias{Storm-class}
\alias{Storm}
\title{Storm  object}
\value{
A S4 object gathering all the following informations
}
\description{
Gather all the needed informations to model a single storm
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the storm}

\item{\code{season}}{numeric. Cyclonic season in which the storm has occured}

\item{\code{basin}}{character. Basin in which the storm has occured}

\item{\code{sshs}}{numeric. Maximum category reached in the Saffir Simpson Hurricane Scale}

\item{\code{numobs.all}}{numeric. Total number of observations available.}

\item{\code{obs.all}}{data.frame. Contains all of the observations available.
An observation is made up of several slots which are:
\itemize{
\item basin
\item subbasin
\item iso.time, Date and hours of observations (UTC)
\item lon, Longitude coordinate (deg)
\item lat, Latitude coordinate (deg)
\item msw, Maximum Sustained Wind (m/s)
\item rmw, Radius of Maximum Wind (km)
\item roci, Radius of the Outermost Closed Isobar (km)
\item pres, pressure at the center (mb)
\item poci, Pressure of the Outermost Closed Isobar (mb)
\item sshs, Category in the Saffir Simpson Hurricane Scale,
}}

\item{\code{numobs}}{numeric. Total number of observations available within the location of interest
extented with its corresponding buffer (See class Storms)}

\item{\code{obs}}{numeric vector. Indices of observations within the location of interest
extented with its corresponding buffer (See class Storms)}

\item{\code{lty.track}}{numeric. Indicates which line type is used to plot the storm on a map}
}}

