% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{stormBehaviour}
\alias{stormBehaviour}
\title{Rasterizing various products among Maximum sustained Wind Speed and PDI,
associated with the each storm}
\usage{
stormBehaviour(
  sts,
  product = "MSW",
  method = "willoughby",
  space_res = 0.1,
  time_res = 1,
  verbose = FALSE,
  threshold = 17
)
}
\arguments{
\item{sts}{Set of storm generated by \code{getStorms} we are interested in}

\item{product}{Characters that represent the product we would like to compute,
currently among \code{MSW} and \code{PDI}.}

\item{method}{Cyclonic model used to compute product. Default value is set to
\code{willoughby}.}

\item{space_res}{Space resolution (in degree) of the incoming raster products.
Default value is set to 0.1 degree.}

\item{time_res}{Time discretization (in hours) to compute the products.
Default value is set to 1h.}

\item{verbose}{Logical, whether or not the function must be verbose. Default
value is set to \code{FALSE}}

\item{threshold}{Represents a wind threshold. If the model provides a value
below this threshold, it is eventually set to \code{NA} as it does not rely on a
tropical storm wind speed.}
}
\value{
A raster stack gathering all the results. Names of the layer are
nameOfTheStorm_product
}
\description{
Rasterizing various products among Maximum sustained Wind Speed and PDI,
associated with the each storm
}
