% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{stormBehaviour}
\alias{stormBehaviour}
\title{Rasterizing various products among Maximum sustained Wind Speed and PDI,
associated with the each storm}
\usage{
stormBehaviour(
  sts,
  product = "MSW",
  method = "Willoughby",
  asymmetry = "None",
  use_rmw = TRUE,
  space_res = 10,
  time_res = 1,
  verbose = FALSE,
  focus_loi = TRUE
)
}
\arguments{
\item{sts}{Set of storm generated by \code{getStorms} we are interested in}

\item{product}{Characters that represent the product we would like to compute,
currently among \code{MSW} and \code{PDI}.}

\item{method}{Cyclonic model used to compute product. Default value is set to
\code{Willoughby}.}

\item{asymmetry}{Character that indicates which asymmetry to use in the computations.
Default value is set to "None".}

\item{use_rmw}{Logical, whether or not using the Radius of maximum wind speed
from the database. Default value is set to TRUE}

\item{space_res}{Space resolution (km) of the incoming raster products.
Default value is set to 10km.}

\item{time_res}{Time discretization (in hours) to compute the products.
Default value is set to 1h.}

\item{verbose}{Logical, whether or not the function must be verbose. Default
value is set to \code{FALSE}}

\item{focus_loi}{Logical, whether or not the computations must only take
place within the loi. Default
value is set to \code{TRUE}}
}
\value{
A raster stack gathering all the results. Names of the layer are
nameOfTheStorm_product
}
\description{
Rasterizing various products among Maximum sustained Wind Speed and PDI,
associated with the each storm
}
