% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{stormBehaviour}
\alias{stormBehaviour}
\title{Rasterizing various products among Maximum sustained Wind Speed and PDI,
associated with the each storm}
\usage{
stormBehaviour(
  sts,
  product = "MSW",
  method = "Willoughby",
  asymmetry = "None",
  empirical_rmw = FALSE,
  result = "analytic",
  space_res = 10,
  time_res = 1,
  verbose = FALSE,
  focus_loi = TRUE
)
}
\arguments{
\item{sts}{Storms object}

\item{product}{characters. Product we would like to compute, that
is either \code{MSW}, (Maximum Sustained Wind) \code{PDI}, (Power Dissipation Index) or \code{Exposure}}

\item{method}{characters. Cyclonic model used to compute product, that is either
\code{Willoughby} and \code{Holland80}.Default value is set to \code{Willoughby}.}

\item{asymmetry}{character. Indicates which asymmetry formula to use in the
computations, that is either \code{None},(no asymmetry) \code{V1} (first version) or
\code{V2} (second version). Default value is set to \code{None}.}

\item{empirical_rmw}{logical. Whether to compute the Radius of Maximum Wind
according to \code{getRmw} or using the Radius of Maximum Wind from the observations.
Default value is set to \code{FALSE}}

\item{result}{...}

\item{space_res}{numeric. Space resolution (km) for the raster to compute.
Default value is set to 10}

\item{time_res}{numeric. Time discretization (hours) used to compute the analytic
Storm rasters. Allowed values are \code{1}, \code{0.75}, \code{0.5}, and \code{0.25}. Default value
is set to 1}

\item{verbose}{logical. Whether or not the function must be verbose and display
a text progress bar. Default value is set to \code{FALSE}}

\item{focus_loi}{logical. Whether or not the computations must only be overtaken
within the \code{spatial.loi.buffer} from \code{sts} object. Default value is set to \code{TRUE},
otherwise, computations are extended over the whole track of the storms}
}
\value{
A raster stack gathering all the results. Names of the layer are
nameOfTheStorm_product
}
\description{
Rasterizing various products among Maximum sustained Wind Speed and PDI,
associated with the each storm
}
