% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{stormBehaviour_pt}
\alias{stormBehaviour_pt}
\title{Compute product on given points coordinates for given storms}
\usage{
stormBehaviour_pt(
  sts,
  points,
  product = "TS",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Boose01",
  empirical_rmw = FALSE,
  time_res = 1
)
}
\arguments{
\item{sts}{Storms object}

\item{points}{data.frame. Contains longitude/latitude coordinates within column
names "lon" and "lat".}

\item{product}{character. Product to compute that is either
\itemize{
\item "MSW", Maximum Sustained Wind
\item "PDI", Power Dissipation Index
\item "Exposure", hours spent for each and all categories together
}
Default value is set to "MSW"}

\item{wind_threshold}{numeric vector. Wind threshold}

\item{method}{character. Cyclonic model used to compute product, that is either
\itemize{
\item "Willoughby"
\item "Holland"
}
Default value is set to "Willoughby"}

\item{asymmetry}{character. Indicates which version of asymmetry to use in
the computations, that is either
\itemize{
\item "None", no asymmetry
\item "Boose01", first version
}
Default value is set to "None".}

\item{empirical_rmw}{logical. Whether to compute the Radius of Maximum Wind
empirically, according to getRmw function or using the Radius of Maximum Wind
from the observations. Default value is set to FALSE}

\item{time_res}{numeric. Time discretization (hours) used to compute the
analytic raster(s). Allowed values are 1 (60min), 0.75 (45min), 0.5
(30min), and 0.25 (15min). Default value is set to 1}
}
\value{
Computed product for each points are returned
through a named list. Each slot is named after the storm and has dimensions:
\itemize{
\item (number of observations:number of point coordinates), if product == "MSW".
Correspond in fact of time series of wind speed on every observations (real and interpolated)
\item (1:number of point coordinates), if product == "PDI"
\item (6:number of point coordinates), if product == "Exposure"
}
}
\description{
This function computes products for each storm of a \code{Storms} object
among TS (time series of wind speed), Power Dissipation Index and Category exposure for each
given points provided in points input
}
