% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{Storms-class}
\alias{Storms-class}
\alias{Storms}
\title{Models a set of storms that occured in a location of interest}
\value{
A S4 object gathering all the above informations
}
\description{
Models a set of storms that occured in a location of interest
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of S4 Storm we are interested in}

\item{\code{time.period}}{Cyclonic season(s) we are interested in}

\item{\code{names}}{Storms names of interest}

\item{\code{sshs}}{numeric. Category in the sshs scale}

\item{\code{nb.storms}}{Number of storms contained in this object}

\item{\code{basin}}{in which basin the TC has occured}

\item{\code{spatial.loi}}{A SpatialPolygons that represents the location of interest.
Projection is EPSG:4326}

\item{\code{buffer}}{the buffer used to extent \code{spatial.loi}}

\item{\code{spatial.loi.buffer}}{buffer extension of \code{spatial.loi}}
}}

