% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{Storms-class}
\alias{Storms-class}
\alias{Storms}
\title{Models a set of storms that occured in a location of interest}
\value{
A S4 object gathering all the above informations
}
\description{
Models a set of storms that occured in a location of interest
}
\section{Slots}{

\describe{
\item{\code{storms}}{A list of S4 Storm we are interested in}

\item{\code{spatial.loi}}{A SpatialPolygons that represents the location of interest.
Projection is EPSG:4326}

\item{\code{aoi}}{A numeric vector that contains longitude/latitude coordinate we are
interested in (if no loi is specified)}
}}

