% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{Storms-class}
\alias{Storms-class}
\alias{Storms}
\title{Storms object}
\value{
A S4 object gathering all the following informations
}
\description{
Gather all the needed informations to model a set of storms
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of Storm objects}

\item{\code{time.period}}{numeric vector. (Range of the) cyclonic seasons of Storms available
in \code{data}}

\item{\code{names}}{character vector. Names of Storms available in data}

\item{\code{sshs}}{numeric vector. Maximum category reached in the Saffir Simpson Hurricane Scale
for all storms available in data}

\item{\code{nb.storms}}{numeric. Total number of storms available in data}

\item{\code{basin}}{character. Basin in which the Storms have occured}

\item{\code{loi.basin}}{logical. Whether the Location Of Interest represents the whole basin or not}

\item{\code{spatial.loi}}{sf object. Represents the location of interest. Projection is EPSG:4326}

\item{\code{buffer}}{numeric. Buffer used to extent spatial.loi (km)}

\item{\code{spatial.loi.buffer}}{buffer extension of spatial.loi}
}}

