% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\name{Storms}
\alias{Storms}
\title{Creating a \code{StormsList} object}
\usage{
Storms(
  sds = test_dataset,
  loi,
  seasons = c(sds@seasons["min"], sds@seasons["max"]),
  names = NULL,
  max_dist = 300,
  remove_TD = TRUE,
  verbose = 2
)
}
\arguments{
\item{sds}{A \code{StormsDataset} object. By default the \code{test_dataset} object is used}

\item{loi}{Location of interest. Can be defined using:
\itemize{
\item A country name (a character object, e.g., "Vanuatu").
\item A point coordinate (a numeric vector with lon, lat in decimal degrees, e.g., c(169.5, -19.2)).
\item A sp (SpatialPolygon) or a sf (simple features) object (e.g., created from a shapefile).
}}

\item{seasons}{Season(s) of occurrence of the storms. A numeric vector of one (e.g., 2020)
or two years (e.g., c(2020,2022)) to extract storms occurring during one specific season or
during several consecutive seasons, respectively. By default all storms occurring since 1980 are extracted.}

\item{names}{Names of specific storms (in capital letters, character vector) to extract.}

\item{max_dist}{Maximum distance between the location of interest and the storm for which track data are extracted (numeric).
Default \code{max_dist} is set to 300 km.}

\item{remove_TD}{Whether (TRUE) or not (FALSE) removing tropical depressions (msw < 18
m/s). Default value is set to \code{TRUE}.}

\item{verbose}{Type of information the function displays (numeric):
\itemize{
\item \code{0}: Nothing is displayed.
\item \code{1}: Only information about the processes are displayed.
\item \code{2}: Information about both the processes and the outputs are displayed (default value).
}}
}
\value{
The \code{Storms()} function a \code{StormsList} object containing track data for all storms
meeting the specified criteria (name, season, location, ...).
}
\description{
The \code{Storms()} function extracts storm track data from a \code{StormsDataset}
and creates a \code{StormsList} object.
}
\details{
The available countries for the \code{loi} are those provided in the
\code{rwolrdxtra} package. This package provide high resolution vector country
boundaries derived from Natural Earth data. More informations on the Natural Earth data
here: http://www.naturalearthdata.com/downloads/10m-cultural-vectors/.
}
\examples{
\dontrun{
#Getting data using country names
vanuatu.st <- Storms(loi = "Vanuatu")

#Getting data using a specific point location
pt <- c(169, -19)
pam.pt <- Storms(loi = pt, names = "PAM")

#Getting data using country and storm names
niran.nc <- Storms(loi = "New Caledonia", names = c("NIRAN"))

#Getting data using a user defined spatial polygon
poly <- cbind(c(135, 290, 290, 135, 135),c(-60, -60, 0, 0, -60))
sp <- sf::st_polygon(list(poly))
sp <- sf::st_sfc(sp, crs = 4326)
sp <- sf::st_as_sf(sp)
sts_sp <- Storms(loi = sp)
}


}
