% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsDataset.R
\docType{class}
\name{StormsDataset-class}
\alias{StormsDataset-class}
\alias{StormsDataset}
\title{StormsDataset object}
\description{
Choose the database to use within the package's functions
}
\section{Slots}{

\describe{
\item{\code{filename}}{character. Name of the file to be loaded}

\item{\code{fields}}{named character vector. Dictionary that provides all the name of dimension to extract from the netcdf database.
See Details}

\item{\code{basin}}{character. Basin name to filter the database within its boundaries. If NULL, no filter is performed and the
whole database will be collected when calling collectData method.}

\item{\code{database}}{list of 12 slots (See Details)}

\item{\code{loaded}}{logical. Whether or not the database have already been loaded. If TRUE, it cannot be loaded anymore}
}}

