% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStorms.R
\docType{class}
\name{StormsList-class}
\alias{StormsList-class}
\alias{StormsList}
\title{StormsList object}
\description{
Gather all the needed informations to model a set of storms
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of Storm objects (See Storm class)}

\item{\code{buffer}}{numeric. Buffer used to extent spatial.loi (km)}

\item{\code{spatial.loi}}{sf object. Represents the location of interest. Projection is EPSG:4326}

\item{\code{spatial.loi.buffer}}{sf object. Buffer extension of spatial.loi}
}}

