% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defStormsList.R
\docType{class}
\name{stormsList-class}
\alias{stormsList-class}
\alias{stormsList}
\title{\code{stormsList} object}
\value{
A \code{stormsList} object.
\itemize{
\item \code{data}, list.
\item \code{buffer}, numeric.
\item \code{spatialLoi}, sf.
\item \code{spatialLoiBuffer}, sf.
}
}
\description{
Gather all the needed informations to model a set of storms
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of \code{storm} (See \code{storm} class)}

\item{\code{buffer}}{numeric. Buffer used to extent \code{spatialLoi} (km)}

\item{\code{spatialLoi}}{sf object. Represents the location of interest. Projection
is EPSG:4326}

\item{\code{spatialLoiBuffer}}{sf object. Buffer extension of \code{spatialLoi}}
}}

