% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormBehaviour.R
\name{temporalBehaviour}
\alias{temporalBehaviour}
\title{Compute indicators of storm behaviour}
\usage{
temporalBehaviour(
  sts,
  points,
  product = "TS",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Chen",
  empirical_rmw = FALSE,
  temp_res = 1,
  verbose = 1
)
}
\arguments{
\item{sts}{\code{StormsList} object.}

\item{points}{data.frame. The lon (in a \code{x} col), lat (in a \code{y} col) coordinates in decimal degrees.}

\item{product}{character. Desired output among:
\itemize{
\item \code{"TS"}: time series of wind speed and direction (default setting).
\item \code{"PDI"}: power dissipation index.
\item \code{"Exposure"}: the duration of exposure.
}}

\item{wind_threshold}{numeric vector. Minimal wind threshold(s) (in $m.s^{-1}$) used to
compute the duration of exposure when \code{product="Exposure"}. By default value the thresholds
used in the Saffir-Simpson hurricane wind scale are used (i.e., 18, 33, 42, 49, 58, 70 $m.s^{-1}$).}

\item{method}{character. Model used to compute wind speed and direction.
Three different models are implemented:
\itemize{
\item \code{"Willoughby"}: the symmetrical model developed by Willoughby et al. (2006) (default setting).
\item \code{"Holland"}: the symmetrical model developed by Holland (1980).
\item \code{"Boose"}: the asymmetrical model developed by  Boose et al. (2004).
}}

\item{asymmetry}{character. If \code{method="Holland"} or \code{method="Willoughby"},
which method is used to add asymmetry. Can be:
\itemize{
\item \code{"Chen"}: the model developed by Chen (1994) (default setting).
\item \code{"Miyazaki"}: the model developed by Miyazaki et al. (1962).
\item \code{"None"}: no asymmetry.
}}

\item{empirical_rmw}{logical. Whether (TRUE) or not (FALSE) to compute
the radius of maximum wind (\code{rmw}) empirically with the model developed by
Willoughby et al. (2006). If \code{empirical_rmw==FALSE} (default setting) then the
\code{rmw} provided in the \code{StormsList} is used.}

\item{temp_res}{numeric. Temporal resolution. Can be \code{1} (for 60 min, default setting),
\code{0.75} (for 45min), \code{0.5} (for 30 min), and \code{0.25} (15 for min).}

\item{verbose}{numeric. Information displayed. Can be:
\itemize{
\item \code{2}: information about the processes and outputs are displayed (default setting).
\item \code{1}: information about the processes are displayed.
\item \code{0}: no information displayed.
}}
}
\value{
For each storm and each point location, the \code{temporalBehaviour()} function returns
a data.frame. The data frames are organised into named lists. Depending on the \code{product} argument
different data.frame are returned:
\itemize{
\item If \code{product == "TS"}, the function returns a data.frame with
one row for each observation (or interpolated observation) and
four columns for wind speed (in $m.s^{-1}$), wind direction (in degree),
the indices, and the ISO time of observations.
\item If \code{product == "PDI"}, the function returns a data.frame with one row
for each point location and one column for the PDI.
\item If \code{product == "Exposure"}, the function returns a data.frame with one
row for each wind speed threshold defined with the \code{wind_threshold} argument
and one column for each point location.
}
}
\description{
The \code{temporalBehaviour()} function allows computing wind speed and direction
for a given location or set of locations along the lifespan of a tropical cyclone.
It also allows to compute to compute three associated summary statistics.
}
\details{
The \code{temp_res} input will perform a linear interpolation of
observations to further compute each 2D wind speed structure at each
interpolated observations. For example, if \code{temp_res == 1}, it will
generate observations every 1 hour between the available observations.
}
\examples{
\dontrun{
pts <- data.frame(x = c(166.5, 163), y = c(-22, -19))
row.names(pts)<-c("point_1","point_2")

#Computing time series of wind speed and direction for Erica and Niran
#over points 1 and 2 defined above
ts_nc <- temporalBehaviour(sts_nc, points = pts)

#Computing PDI for Erica and Niran over points 1 and 2 defined above
pdiPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "PDI")

#Computing the duration of exposure to wind speeds above the thresholds
#used by the Saffir-Simpson hurricane wind scale for Erica and Niran
#over points 1 and 2 defined above
expPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "Exposure", wind_threshold = c(20,30))
}

}
\references{
Boose, E. R., Serrano, M. I., & Foster, D. R. (2004). Landscape and regional impacts of hurricanes in Puerto Rico.
Ecological Monographs, 74(2), Article 2. https://doi.org/10.1890/02-4057

Chen, K.-M. (1994). A computation method for typhoon wind field. Tropic Oceanology, 13(2), 41–48.

Holland, G. J. (1980). An Analytic Model of the Wind and Pressure Profiles in Hurricanes. Monthly Weather Review, 108(8), 1212–1218.
https://doi.org/10.1175/1520-0493(1980)108<1212:AAMOTW>2.0.CO;2

Miyazaki, M., Ueno, T., & Unoki, S. (1962). The theoretical investigations of typhoon surges along the Japanese coast (II).
Oceanographical Magazine, 13(2), 103–117.

Willoughby, H. E., Darling, R. W. R., & Rahn, M. E. (2006). Parametric Representation of the Primary Hurricane Vortex.
Part II: A New Family of Sectionally Continuous Profiles. Monthly Weather Review, 134(4), 1102–1120. https://doi.org/10.1175/MWR3106.1
}
