% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRast.R
\name{writeRast}
\alias{writeRast}
\title{Save raster(s) in the desired format}
\usage{
writeRast(rast, format = ".tiff", filename = NULL, path = "./")
}
\arguments{
\item{rast}{SpatRaster object}

\item{format}{character. Format of the file to export. Either \code{".tiff"} or
\code{".nc"}. Default value is set to \code{".tiff"}}

\item{filename}{character. Name of the file. Default value is set to \code{NULL},
in this case it will be set to \code{names(rast)}}

\item{path}{character. Relative path where the file should be written}
}
\value{
\code{NULL}
}
\description{
This function writes SpatRaster(s) in the given format among Geotiff or
netcdf
}
\examples{
\dontrun{
pam <- Storms(loi = "Vanuatu", names = "PAM")
pam.msw <- spatialBehaviour(pam)
#Save MSW raster in Geotiff file for Pam 2015 over Vanuatu
writeRast(pam.msw, path = paste0(tempdir(),"/"))

sts.nc <- Storms(loi = "New Caledonia")
pdi.nc <- spatialBehaviour(sts.nc, product = "PDI")
#Save PDI rasters in Geotiff files contained in pdi.nc
writeRast(pdi.nc, path = paste0(tempdir(),"/"))
}

}
