% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRast.R
\name{writeRast}
\alias{writeRast}
\title{Write a SpatRast raster in the given format}
\usage{
writeRast(rast, format = ".tiff", filename = "rast", path = "./")
}
\arguments{
\item{rast}{SpatRast object we wish to write}

\item{format}{format of the file we want to export the \code{rast}}

\item{filename}{name of the file}

\item{path}{the relative path where the file shoule be written}
}
\description{
Write a SpatRast raster in the given format
}
