% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRast.R
\name{writeRast}
\alias{writeRast}
\title{Save raster(s) in the desired format}
\usage{
writeRast(rast, format = ".tiff", filename = NULL, path = "./")
}
\arguments{
\item{rast}{SpatRaster object}

\item{format}{character. Format of the file to export. Either ".tiff"
or ".nc". Default value is set to ".tiff"}

\item{filename}{character. Name of the file. Default value is set to NULL,
in this case it will be set to names(rast)}

\item{path}{character. Relative path where the file should be written}
}
\description{
This function writes SpatRaster(s) in the given format among Geotiff or
netcdf
}
\examples{
\dontrun{
#Save MSW raster in Geotiff file for Pam 2015 over Vanuatu
pam <- getStorms(loi = "Vanuatu", names = "PAM")
pam.msw <- stormBehaviour_sp(pam)
writeRast(pam.msw, path = paste0(tempdir(),"/"))

#Save PDI rasters in Geotiff files for Erica 2003 and Niran 2021 over
#New Caledonia
sts.nc <- getStorms(loi = "New Caledonia", names = c("ERICA", "NIRAN"))
pdi.nc <- stormbehaviour_sp(sts.nc, product = "PDI")
writeRast(pdi.nc, path = paste0(tempdir(),"/"))
}

}
