% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRast.R
\name{writeRast}
\alias{writeRast}
\title{Save raster(s) in the desired format}
\usage{
writeRast(rast, format = ".tiff", filename = NULL, path = "./")
}
\arguments{
\item{rast}{SpatRaster object}

\item{format}{character. Format of the file to export. Either ".tiff"
or ".nc". Default value is set to ".tiff"}

\item{filename}{character. Name of the file. Default value is set to NULL,
in this case it will be set to names(rast)}

\item{path}{character. Relative path where the file should be written}
}
\description{
This function writes SpatRaster(s) in the given format among Geotiff or
netcdf
}
\examples{
#Save MSW raster in Geotiff file for PAM (2015) over Vanuatu
pam_msw <- terra::rast(system.file("extdata", "PAM_MSW.tiff",
                                  package = "StormR"))
writeRast(pam_msw, path = paste0(tempdir(),"/"))

#Save PDI rasters in Geotiff files for ERICA (2003) and NIRAN (2015)
#over New Caledonia
erica_pdi <- terra::rast(system.file("extdata", "ERICA_PDI.tiff",
                                    package = "StormR"))
niran_pdi <- terra::rast(system.file("extdata", "NIRAN_PDI.tiff",
                                    package = "StormR"))
pdi_nc <- c(erica_pdi,niran_pdi)
writeRast(pdi_nc, path = paste0(tempdir(),"/"))

}
