% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRast.R
\name{writeRast}
\alias{writeRast}
\title{Exporting rasters to GeoTIFF or NetCDF files}
\usage{
writeRast(rast, format = ".tiff", filename = NULL, path = "./")
}
\arguments{
\item{rast}{\code{SpatRaster} object.}

\item{format}{character. Format of the file to export. Either \code{".tiff"} (GeoTIFF, default setting) or
\code{".nc"} (NetCDF).}

\item{filename}{character. Output file name. By default \code{filename=NULL} and the name of the raster layer is used.}

\item{path}{character. Path to the directory where the file is exported to.}
}
\value{
\code{NULL}
}
\description{
The \code{writeRast()} function exports rasters stored in SpatRaster objects to
GeoTIFF or NetCDF files.
}
\examples{
\dontrun{
#Creating a StormsDataset
sds <- defDatabase()

#Getting storm track data for tropical cyclone Pam (2015) near Vanuatu
pam <- Storms(sds = sds, loi = "Vanuatu", names = "PAM")

#Computing maximum sustained wind speed
pam.msw <- spatialBehaviour(pam)

#Exporting maximum sustained wind speed raster layer to a GeoTIFF file
writeRast(pam.msw, path = paste0(tempdir(),"/"))

#Computing power dissipation index for several storms near New Caledonia
sts.nc <- Storms(sds = sds, loi = "New Caledonia")
pdi.nc <- spatialBehaviour(sts.nc, product = "PDI")

#Exporting the power dissipation index raster layers to a NetCDF file
writeRast(pdi.nc, path = paste0(tempdir(),"/"))
}

}
